% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Calc_Change}
\alias{Calc_Change}
\title{Utility function to calculate the change to make each iteration
\code{Calc_Change} Called to update the parameter changes, Uses log-likelihoods and control parameters, Applys newton steps and change limitations}
\usage{
Calc_Change(
  nthreads,
  totalnum,
  fir,
  der_iden,
  dbeta_cap,
  dose_abs_max,
  lr,
  abs_max,
  Ll,
  Lld,
  Lldd,
  dbeta,
  change_all,
  tform,
  dint,
  KeepConstant,
  debugging
)
}
\arguments{
\item{nthreads}{number of threads}

\item{totalnum}{total number of parameter}

\item{fir}{first term number}

\item{der_iden}{subterm number for derivative tests}

\item{dbeta_cap}{learning rate for newton step toward 0 log-likelihood}

\item{dose_abs_max}{Maximum allowed threshold parameter change}

\item{lr}{learning rate fo newton step toward 0 derivative}

\item{abs_max}{Maximum allowed parameter change}

\item{Ll}{Log-Likelihood}

\item{Lld}{Log-Likelihood first derivative}

\item{Lldd}{Log-Likelihood second derivative}

\item{dbeta}{parameter change vector}

\item{change_all}{boolean to change every parameter}

\item{tform}{subterm type}

\item{dint}{value used for threshold derivative calculation}

\item{KeepConstant}{vector of parameters to keep constant}

\item{debugging}{debugging boolean}
}
\value{
Updates matrices in place: parameter change matrix
}
\description{
Utility function to calculate the change to make each iteration
\code{Calc_Change} Called to update the parameter changes, Uses log-likelihoods and control parameters, Applys newton steps and change limitations
}
