% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Calc_LogLik}
\alias{Calc_LogLik}
\title{Utility function to calculate Cox Log-Likelihood and derivatives
\code{Calc_LogLik} Called to update log-likelihoods, Uses list of event rows, risk matrices, and repeated sums, Sums the log-likelihood contribution from each event time}
\usage{
Calc_LogLik(
  nthreads,
  RiskFail,
  RiskGroup,
  totalnum,
  ntime,
  R,
  Rd,
  Rdd,
  RdR,
  RddR,
  Rls1,
  Rls2,
  Rls3,
  Lls1,
  Lls2,
  Lls3,
  Ll,
  Lld,
  Lldd,
  debugging,
  ties_method
)
}
\arguments{
\item{nthreads}{number of threads}

\item{RiskFail}{Matrix of event rows for each event time}

\item{RiskGroup}{vectors of strings with rows at risk for each event time}

\item{totalnum}{total number of parameters}

\item{ntime}{number of event times}

\item{R}{Risk matrix}

\item{Rd}{Risk derivative matrix}

\item{Rdd}{Risk second derivative matrix}

\item{RdR}{Risk to first derivative ratio matrix}

\item{RddR}{Risk to second derivative ratio matrix}

\item{Rls1}{First Risk sum storage}

\item{Rls2}{First Risk sum derivative storage}

\item{Rls3}{First Risk sum second derivative storage}

\item{Lls1}{Second Risk sum storage}

\item{Lls2}{Second Risk sum derivative storage}

\item{Lls3}{Second Risk sum second derivative storage}

\item{Ll}{Log-likelihood vector}

\item{Lld}{Log-likelihood first derivative vector}

\item{Lldd}{Log-likelihood second derivative matrix}

\item{debugging}{debugging boolean}

\item{ties_method}{Ties method}
}
\value{
Updates matrices in place: Log-likelihood vectors/matrix
}
\description{
Utility function to calculate Cox Log-Likelihood and derivatives
\code{Calc_LogLik} Called to update log-likelihoods, Uses list of event rows, risk matrices, and repeated sums, Sums the log-likelihood contribution from each event time
}
