% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Calc_LogLik_STRATA_SINGLE}
\alias{Calc_LogLik_STRATA_SINGLE}
\title{Utility function to calculate just Cox Log-Likelihood with STRATA
\code{Calc_LogLik_STRATA_SINGLE} Called to update log-likelihoods, Uses list of event rows, risk matrices, and repeated sums, Sums the log-likelihood contribution from each event time and strata}
\usage{
Calc_LogLik_STRATA_SINGLE(
  nthreads,
  RiskFail,
  RiskGroup,
  totalnum,
  ntime,
  R,
  Rls1,
  Lls1,
  Ll,
  debugging,
  ties_method,
  STRATA_vals,
  KeepConstant
)
}
\arguments{
\item{nthreads}{number of threads}

\item{RiskFail}{Matrix of event rows for each event time}

\item{RiskGroup}{vectors of strings with rows at risk for each event time}

\item{totalnum}{total number of parameters}

\item{ntime}{number of event times}

\item{R}{Risk matrix}

\item{Rls1}{First Risk sum storage}

\item{Lls1}{Second Risk sum storage}

\item{Ll}{Log-likelihood vector}

\item{debugging}{debugging boolean}

\item{ties_method}{Ties method}

\item{STRATA_vals}{vector of strata identifier values}

\item{KeepConstant}{vector identifying constant parameters}
}
\value{
Updates matrices in place: Log-likelihood vectors/matrix
}
\description{
Utility function to calculate just Cox Log-Likelihood with STRATA
\code{Calc_LogLik_STRATA_SINGLE} Called to update log-likelihoods, Uses list of event rows, risk matrices, and repeated sums, Sums the log-likelihood contribution from each event time and strata
}
