% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Calc_Null_LogLik}
\alias{Calc_Null_LogLik}
\title{Utility function to perform null model equivalent of Calc_LogLik
\code{Calc_Null_LogLik} Called to update log-likelihoods, Uses list of event rows and repeated sums, Sums the log-likelihood contribution from each event time}
\usage{
Calc_Null_LogLik(
  nthreads,
  RiskFail,
  RiskGroup,
  ntime,
  R,
  Rls1,
  Lls1,
  Ll,
  ties_method
)
}
\arguments{
\item{nthreads}{number of threads}

\item{RiskFail}{Matrix of event rows for each event time}

\item{RiskGroup}{vectors of strings with rows at risk for each event time}

\item{ntime}{number of event times}

\item{R}{Risk matrix}

\item{Rls1}{First Risk sum storage}

\item{Lls1}{Second Risk sum storage}

\item{Ll}{Log-likelihood vector}

\item{ties_method}{Ties method}
}
\value{
Updates matrices in place: Log-likelihood vectors/matrix
}
\description{
Utility function to perform null model equivalent of Calc_LogLik
\code{Calc_Null_LogLik} Called to update log-likelihoods, Uses list of event rows and repeated sums, Sums the log-likelihood contribution from each event time
}
