% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Calculate_Null_Sides}
\alias{Calculate_Null_Sides}
\title{Utility function to perform null model equivalent of Calculate_Sides
\code{Calculate_Null_Sides} Called to update repeated sum calculations, Uses list of event rows, Performs calculation of counts in each group}
\usage{
Calculate_Null_Sides(RiskFail, RiskGroup, ntime, R, Rls1, Lls1, nthreads)
}
\arguments{
\item{RiskFail}{Matrix of event rows for each event time}

\item{RiskGroup}{vectors of strings with rows at risk for each event time}

\item{ntime}{number of event times}

\item{R}{Risk matrix}

\item{Rls1}{First Risk sum storage}

\item{Lls1}{Second Risk sum storage}

\item{nthreads}{number of threads}
}
\value{
Updates matrices in place: risk storage matrices
}
\description{
Utility function to perform null model equivalent of Calculate_Sides
\code{Calculate_Null_Sides} Called to update repeated sum calculations, Uses list of event rows, Performs calculation of counts in each group
}
