% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Calculate_Null_Sides_STRATA}
\alias{Calculate_Null_Sides_STRATA}
\title{Utility function to perform null model equivalent of Calculate_Sides with strata
\code{Calculate_Null_Sides_STRATA} Called to update repeated sum calculations, Uses list of event rows, Performs calculation of counts in each group}
\usage{
Calculate_Null_Sides_STRATA(
  RiskFail,
  RiskGroup,
  ntime,
  R,
  Rls1,
  Lls1,
  STRATA_vals,
  nthreads
)
}
\arguments{
\item{RiskFail}{matrix of indexes for event rows in each risk group}

\item{RiskGroup}{list of string indices for every risk group}

\item{ntime}{number of risk groups}

\item{R}{risk vector}

\item{Rls1}{First Risk sum storage}

\item{Lls1}{Second Risk sum storage}

\item{STRATA_vals}{vector of strata identifier values}

\item{nthreads}{number of threads available}
}
\value{
Updates matrices in place: risk storage matrices
}
\description{
Utility function to perform null model equivalent of Calculate_Sides with strata
\code{Calculate_Null_Sides_STRATA} Called to update repeated sum calculations, Uses list of event rows, Performs calculation of counts in each group
}
