% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Calculate_Sides_Single}
\alias{Calculate_Sides_Single}
\title{Utility function to calculate repeated values used in Cox Log-Likelihood calculation. but not derivatives
\code{Calculate_Sides_Single} Called to update repeated sum calculations, Uses list of event rows and risk matrices, Performs calculation of sums of risk in each group}
\usage{
Calculate_Sides_Single(
  RiskFail,
  RiskGroup,
  totalnum,
  ntime,
  R,
  Rls1,
  Lls1,
  nthreads,
  debugging
)
}
\arguments{
\item{RiskFail}{Matrix of event rows for each event time}

\item{RiskGroup}{vectors of strings with rows at risk for each event time}

\item{totalnum}{total number of parameters}

\item{ntime}{number of event times}

\item{R}{Risk matrix}

\item{Rls1}{First Risk sum storage}

\item{Lls1}{Second Risk sum storage}

\item{nthreads}{number of threads}

\item{debugging}{debugging boolean}
}
\value{
Updates matrices in place: risk storage matrices
}
\description{
Utility function to calculate repeated values used in Cox Log-Likelihood calculation. but not derivatives
\code{Calculate_Sides_Single} Called to update repeated sum calculations, Uses list of event rows and risk matrices, Performs calculation of sums of risk in each group
}
