% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Calculate_Sides_STRATA}
\alias{Calculate_Sides_STRATA}
\title{Utility function to calculate repeated values used in Cox Log-Likelihood calculation with STRATA
\code{Calculate_Sides_STRATA} Called to update repeated sum calculations, Uses list of event rows and risk matrices, Performs calculation of sums of risk in each group}
\usage{
Calculate_Sides_STRATA(
  RiskFail,
  RiskGroup,
  totalnum,
  ntime,
  R,
  Rd,
  Rdd,
  Rls1,
  Rls2,
  Rls3,
  Lls1,
  Lls2,
  Lls3,
  nthreads,
  debugging,
  STRATA_vals,
  KeepConstant
)
}
\arguments{
\item{RiskFail}{Matrix of event rows for each event time}

\item{RiskGroup}{vectors of strings with rows at risk for each event time}

\item{totalnum}{total number of parameters}

\item{ntime}{number of event times}

\item{R}{Risk matrix}

\item{Rd}{Risk derivative matrix}

\item{Rdd}{Risk second derivative matrix}

\item{Rls1}{First Risk sum storage}

\item{Rls2}{First Risk sum derivative storage}

\item{Rls3}{First Risk sum second derivative storage}

\item{Lls1}{Second Risk sum storage}

\item{Lls2}{Second Risk sum derivative storage}

\item{Lls3}{Second Risk sum second derivative storage}

\item{nthreads}{number of threads}

\item{debugging}{debugging boolean}

\item{STRATA_vals}{vector of strata identifier values}

\item{KeepConstant}{vector identifying constant parameters}
}
\value{
Updates matrices in place: risk storage matrices
}
\description{
Utility function to calculate repeated values used in Cox Log-Likelihood calculation with STRATA
\code{Calculate_Sides_STRATA} Called to update repeated sum calculations, Uses list of event rows and risk matrices, Performs calculation of sums of risk in each group
}
