% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Check_Cluster}
\alias{Check_Cluster}
\title{Checks for clusters of scores
\code{Check_Cluster} generates the center of a cluster}
\usage{
Check_Cluster(df, names, dev_column)
}
\arguments{
\item{df}{dataframe of data to use as reference}

\item{names}{list of columns names}

\item{dev_column}{column with scores}
}
\value{
returns the updated dataframe
}
\description{
Checks for clusters of scores
\code{Check_Cluster} generates the center of a cluster
}
