% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Check_Iters}
\alias{Check_Iters}
\title{Automatically checks the number of starting guesses}
\usage{
Check_Iters(control, a_n)
}
\arguments{
\item{control}{list of parameters controlling the convergence, see the Control_Options vignette for details}

\item{a_n}{list of initial parameter values, used to determine the number of parameters. May be either a list of vectors or a single vector.}
}
\value{
returns a list with the corrected control list and a_n
}
\description{
\code{Check_Iters} checks the number of iterations and number of guesses, and corrects
}
\seealso{
Other Data Cleaning Functions: 
\code{\link{Date_Shift}()},
\code{\link{Event_Count_Gen}()},
\code{\link{Event_Time_Gen}()},
\code{\link{Joint_Multiple_Events}()},
\code{\link{Replace_Missing}()},
\code{\link{Time_Since}()},
\code{\link{apply_norm}()},
\code{\link{factorize}()},
\code{\link{gen_time_dep}()}
}
\concept{Data Cleaning Functions}
