% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Check_Risk}
\alias{Check_Risk}
\title{checks if the model is viable
\code{Check_Risk} Calculates risks and checks for negative values}
\usage{
Check_Risk(
  Term_n,
  tform,
  a_n,
  x_all,
  dfc,
  fir,
  modelform,
  verbose,
  debugging,
  KeepConstant,
  term_tot,
  nthreads
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{x_all}{covariate matrix}

\item{dfc}{covariate column numbers}

\item{fir}{first term number}

\item{modelform}{model string}

\item{verbose}{verbosity boolean}

\item{debugging}{debugging boolean}

\item{KeepConstant}{vector identifying constant parameters}

\item{term_tot}{total number of terms}

\item{nthreads}{number of threads to use}
}
\value{
True for viable point, False for negative error
}
\description{
checks if the model is viable
\code{Check_Risk} Calculates risks and checks for negative values
}
