% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Check_Verbose}
\alias{Check_Verbose}
\title{General purpose verbosity check}
\usage{
Check_Verbose(verbose)
}
\arguments{
\item{verbose}{integer valued 0-4 controlling what information is printed to the terminal. Each level includes the lower levels. 0: silent, 1: errors printed, 2: warnings printed, 3: notes printed, 4: debug information printed. Errors are situations that stop the regression, warnings are situations that assume default values that the user might not have intended, notes provide information on regression progress, and debug prints out C++ progress and intermediate results. The default level is 2 and True/False is converted to 3/0.}
}
\value{
returns correct verbose value
}
\description{
\code{Check_Verbose} checks and assigns verbosity values
}
\seealso{
Other Data Cleaning Functions: 
\code{\link{Check_Dupe_Columns}()},
\code{\link{Check_Trunc}()},
\code{\link{Date_Shift}()},
\code{\link{Def_Control}()},
\code{\link{Def_model_control}()},
\code{\link{Event_Count_Gen}()},
\code{\link{Event_Time_Gen}()},
\code{\link{Joint_Multiple_Events}()},
\code{\link{Replace_Missing}()},
\code{\link{Time_Since}()},
\code{\link{factorize}()},
\code{\link{gen_time_dep}()},
\code{\link{interact_them}()},
\code{\link{nested_split}()},
\code{\link{parse_literal_string}()}
}
\concept{Data Cleaning Functions}
