% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{ColossusCaseConStrataSurv}
\alias{ColossusCaseConStrataSurv}
\title{Interprets basic case-control survival formula RHS with grouping by strata}
\usage{
ColossusCaseConStrataSurv(...)
}
\arguments{
\item{...}{entries for a Case-Control object, with strata. Expects an event column and strata column, either named or in order.}
}
\value{
returns list with event
}
\description{
\code{ColossusCaseConStrataSurv} assigns and interprets interval columns for case-control model with grouping by strata.
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCaseConSurv}()},
\code{\link{ColossusCaseConTimeStrataSurv}()},
\code{\link{ColossusCaseConTimeSurv}()},
\code{\link{ColossusCoxStrataSurv}()},
\code{\link{ColossusCoxSurv}()},
\code{\link{ColossusExpressionCall}()},
\code{\link{ColossusFineGrayStrataSurv}()},
\code{\link{ColossusFineGraySurv}()},
\code{\link{ColossusLogitSurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()},
\code{\link{get_form_joint}()},
\code{\link{get_form_list}()},
\code{\link{get_form_risk}()},
\code{\link{get_form_surv}()}
}
\concept{Formula Interpretation}
