% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{ColossusCaseConSurv}
\alias{ColossusCaseConSurv}
\title{Interprets basic case-control survival formula RHS with no grouping}
\usage{
ColossusCaseConSurv(...)
}
\arguments{
\item{...}{entries for a Case-Control object, only the event column}
}
\value{
returns list with event
}
\description{
\code{ColossusCaseConSurv} assigns and interprets interval columns for case-control model without grouping.
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCaseConStrataSurv}()},
\code{\link{ColossusCaseConTimeStrataSurv}()},
\code{\link{ColossusCaseConTimeSurv}()},
\code{\link{ColossusCoxStrataSurv}()},
\code{\link{ColossusCoxSurv}()},
\code{\link{ColossusFineGrayStrataSurv}()},
\code{\link{ColossusFineGraySurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()},
\code{\link{get_form_joint}()},
\code{\link{get_form_list}()},
\code{\link{get_form_risk}()},
\code{\link{get_form_surv}()}
}
\concept{Formula Interpretation}
