% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{ColossusCaseConTimeStrataSurv}
\alias{ColossusCaseConTimeStrataSurv}
\title{Interprets basic case-control survival formula RHS with grouping by strata and risk group}
\usage{
ColossusCaseConTimeStrataSurv(...)
}
\arguments{
\item{...}{entries for a Case-Control object, with strata and interval info. Similar to the stratified Cox survival object. Expects interval start, end, event, and strata in order or named. Removes strata and calls the standard Case-Control survival object grouped by risk group.}
}
\value{
returns list with event
}
\description{
\code{ColossusCaseConTimeStrataSurv} assigns and interprets interval columns for case-control model with grouping by strata and risk group.
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCaseConStrataSurv}()},
\code{\link{ColossusCaseConSurv}()},
\code{\link{ColossusCaseConTimeSurv}()},
\code{\link{ColossusCoxStrataSurv}()},
\code{\link{ColossusCoxSurv}()},
\code{\link{ColossusFineGrayStrataSurv}()},
\code{\link{ColossusFineGraySurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()},
\code{\link{get_form_joint}()},
\code{\link{get_form_list}()},
\code{\link{get_form_risk}()},
\code{\link{get_form_surv}()}
}
\concept{Formula Interpretation}
