% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{ColossusCoxStrataSurv}
\alias{ColossusCoxStrataSurv}
\title{Interprets basic cox survival formula RHS with strata}
\usage{
ColossusCoxStrataSurv(...)
}
\arguments{
\item{...}{entries for a cox survival object with strata, tstart, tend, event, and strata. Either in order or named. If unnamed, the last is assumed to be strata and the rest are passed to ColossusCoxSurv.}
}
\value{
returns list with interval endpoints, event, and strata column
}
\description{
\code{ColossusCoxStrataSurv} assigns and interprets interval columns for cox model with stratification
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCoxSurv}()},
\code{\link{ColossusFineGrayStrataSurv}()},
\code{\link{ColossusFineGraySurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()},
\code{\link{get_form_joint}()},
\code{\link{get_form_list}()},
\code{\link{get_form_risk}()},
\code{\link{get_form_surv}()}
}
\concept{Formula Interpretation}
