% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{ColossusExpressionCall}
\alias{ColossusExpressionCall}
\title{Applies a list of function calls from a model evaluation}
\usage{
ColossusExpressionCall(calls, df)
}
\arguments{
\item{calls}{List from 'model$expres' calls formatted as the type of function call, and then the arguments}

\item{df}{dataset, to be modified by the function calls}
}
\value{
returns the transformed data
}
\description{
\code{ColossusExpressionCall} Uses a list of function call info and a data.table, and applies the function calls
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCaseConStrataSurv}()},
\code{\link{ColossusCaseConSurv}()},
\code{\link{ColossusCaseConTimeStrataSurv}()},
\code{\link{ColossusCaseConTimeSurv}()},
\code{\link{ColossusCoxStrataSurv}()},
\code{\link{ColossusCoxSurv}()},
\code{\link{ColossusFineGrayStrataSurv}()},
\code{\link{ColossusFineGraySurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()},
\code{\link{get_form_joint}()},
\code{\link{get_form_list}()},
\code{\link{get_form_risk}()},
\code{\link{get_form_surv}()}
}
\concept{Formula Interpretation}
