% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{ColossusFineGrayStrataSurv}
\alias{ColossusFineGrayStrataSurv}
\title{Interprets basic fine-gray survival formula RHS with strata}
\usage{
ColossusFineGrayStrataSurv(...)
}
\value{
returns list with interval endpoints, event, strata, and weighting columns
}
\description{
\code{ColossusFineGraySurv} assigns and interprets interval columns for fine-gray model and stratification
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCoxStrataSurv}()},
\code{\link{ColossusCoxSurv}()},
\code{\link{ColossusFineGraySurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()}
}
\concept{Formula Interpretation}
