% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{ColossusFineGraySurv}
\alias{ColossusFineGraySurv}
\title{Interprets basic fine-gray survival formula RHS}
\usage{
ColossusFineGraySurv(...)
}
\value{
returns list with interval endpoints, event, and weighting columns
}
\description{
\code{ColossusFineGraySurv} assigns and interprets interval columns for fine-gray model
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCoxStrataSurv}()},
\code{\link{ColossusCoxSurv}()},
\code{\link{ColossusFineGrayStrataSurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()}
}
\concept{Formula Interpretation}
