% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cox_ph_null}
\alias{cox_ph_null}
\title{Interface between R code and the Cox PH null model function
\code{Stress_Test} Called directly from R, Defines the control variables and calls the calculation function}
\usage{
cox_ph_null(Control, df_groups, tu)
}
\arguments{
\item{Control}{control list}

\item{df_groups}{time and event matrix}

\item{tu}{event times}
}
\value{
LogLike_Cox_PH_null output : Log-likelihood of optimum, AIC
}
\description{
Interface between R code and the Cox PH null model function
\code{Stress_Test} Called directly from R, Defines the control variables and calls the calculation function
}
