% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cox_ph_plot}
\alias{cox_ph_plot}
\title{Interface between R code and the Cox PH plotting
\code{cox_ph_plot} Called directly from R, Defines the control variables and calls the correct plotting function}
\usage{
cox_ph_plot(
  Term_n,
  tform,
  a_n,
  dfc,
  x_all,
  fir,
  der_iden,
  modelform,
  Control,
  df_groups,
  tu,
  KeepConstant,
  term_tot,
  Plot_Type,
  uniq_v
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{dfc}{covariate column numbers}

\item{x_all}{covariate matrix}

\item{fir}{first term number}

\item{der_iden}{subterm number for derivative tests}

\item{modelform}{model string}

\item{Control}{control list}

\item{df_groups}{time and event matrix}

\item{tu}{event times}

\item{KeepConstant}{vector of parameters to keep constant}

\item{term_tot}{total number of terms}

\item{Plot_Type}{string specifying which plot type}

\item{uniq_v}{total number of unique covariate values}
}
\value{
Cox_PH_PLOT_SURV : ( baseline harzard, risk for each row) or Cox_PH_PLOT_RISK output : (covariate values, risks for each row)
}
\description{
Interface between R code and the Cox PH plotting
\code{cox_ph_plot} Called directly from R, Defines the control variables and calls the correct plotting function
}
