% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cox_PH_PLOT_RISK}
\alias{Cox_PH_PLOT_RISK}
\title{Primary Cox PH risk plotting function
\code{Cox_PH_PLOT_RISK} Performs the calls to calculation functions, Uses formula and a generated list of covariate values to calculate risk over a grid, With verbose option prints out time stamps and intermediate sums of terms and derivatives}
\usage{
Cox_PH_PLOT_RISK(
  Term_n,
  tform,
  a_n,
  x_all,
  dfc,
  fir,
  der_iden,
  modelform,
  abs_max,
  dose_abs_max,
  df_groups,
  tu,
  verbose,
  debugging,
  KeepConstant,
  term_tot,
  uniq_v
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{x_all}{covariate matrix}

\item{dfc}{covariate column numbers}

\item{fir}{first term number}

\item{der_iden}{subterm number for derivative tests}

\item{modelform}{model string}

\item{abs_max}{Maximum allowed parameter change}

\item{dose_abs_max}{Maximum allowed threshold parameter change}

\item{df_groups}{matrix with time and event information}

\item{tu}{event times}

\item{verbose}{verbosity boolean}

\item{debugging}{debugging boolean}

\item{KeepConstant}{vector identifying constant parameters}

\item{term_tot}{total number of terms}

\item{uniq_v}{number of unqiue covariate values}
}
\value{
List of results: covariate values, risks for each row
}
\description{
Primary Cox PH risk plotting function
\code{Cox_PH_PLOT_RISK} Performs the calls to calculation functions, Uses formula and a generated list of covariate values to calculate risk over a grid, With verbose option prints out time stamps and intermediate sums of terms and derivatives
}
