% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cox_ph_risk_sub}
\alias{cox_ph_risk_sub}
\title{Interface between R code and the reference risk calculation
\code{cox_ph_risk_sub} Called directly from R, Defines the control variables and calls the calculation function}
\usage{
cox_ph_risk_sub(
  Term_n,
  tform,
  a_n,
  dfc,
  x_all,
  fir,
  modelform,
  Control,
  term_tot
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{dfc}{covariate column numbers}

\item{x_all}{covariate matrix}

\item{fir}{first term number}

\item{modelform}{model string}

\item{Control}{control list}

\item{term_tot}{total number of terms}
}
\value{
RISK_SUBSET output: Risk at the reference
}
\description{
Interface between R code and the reference risk calculation
\code{cox_ph_risk_sub} Called directly from R, Defines the control variables and calls the calculation function
}
