% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cox_ph_STRATA}
\alias{cox_ph_STRATA}
\title{Interface between R code and the Cox PH regression with STRATA
\code{cox_ph_STRATA} Called directly from R, Defines the control variables and calls the regression function}
\usage{
cox_ph_STRATA(
  Term_n,
  tform,
  a_n,
  dfc,
  x_all,
  fir,
  der_iden,
  modelform,
  Control,
  df_groups,
  tu,
  KeepConstant,
  term_tot,
  STRATA_vals
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{dfc}{covariate column numbers}

\item{x_all}{covariate matrix}

\item{fir}{first term number}

\item{der_iden}{subterm number for derivative tests}

\item{modelform}{model string}

\item{Control}{control list}

\item{df_groups}{time and event matrix}

\item{tu}{event times}

\item{KeepConstant}{vector of parameters to keep constant}

\item{term_tot}{total number of terms}

\item{STRATA_vals}{vector of strata identifier values}
}
\value{
LogLik_Cox_PH output : Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, model information
}
\description{
Interface between R code and the Cox PH regression with STRATA
\code{cox_ph_STRATA} Called directly from R, Defines the control variables and calls the regression function
}
