% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cox_ph_transition_basic}
\alias{cox_ph_transition_basic}
\title{Interface between R code and the Cox PH regression for basic model
\code{cox_ph_transition_basic} Called directly from R, Defines the control variables and calls the regression function}
\usage{
cox_ph_transition_basic(
  a_n,
  dfc,
  x_all,
  der_iden,
  Control,
  df_groups,
  tu,
  KeepConstant
)
}
\arguments{
\item{a_n}{starting values}

\item{dfc}{covariate column numbers}

\item{x_all}{covariate matrix}

\item{der_iden}{subterm number for derivative tests}

\item{Control}{control list}

\item{df_groups}{time and event matrix}

\item{tu}{event times}

\item{KeepConstant}{vector of parameters to keep constant}
}
\value{
LogLik_Cox_PH output : Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, model information
}
\description{
Interface between R code and the Cox PH regression for basic model
\code{cox_ph_transition_basic} Called directly from R, Defines the control variables and calls the regression function
}
