% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cox_ph_transition_CR_Single}
\alias{cox_ph_transition_CR_Single}
\title{Interface between R code and the Cox PH calculation with competing risks, event=2
\code{cox_ph_transition_CR_Single} Called directly from R, Defines the control variables and calls the regression function}
\usage{
cox_ph_transition_CR_Single(
  Term_n,
  tform,
  a_n,
  dfc,
  x_all,
  fir,
  modelform,
  Control,
  df_groups,
  tu,
  cens_vec,
  KeepConstant,
  term_tot
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{dfc}{covariate column numbers}

\item{x_all}{covariate matrix}

\item{fir}{first term number}

\item{modelform}{model string}

\item{Control}{control list}

\item{df_groups}{time and event matrix}

\item{tu}{event times}

\item{cens_vec}{censoring weight list}

\item{KeepConstant}{vector of parameters to keep constant}

\item{term_tot}{total number of terms}
}
\value{
LogLik_Cox_PH output : Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, model information
}
\description{
Interface between R code and the Cox PH calculation with competing risks, event=2
\code{cox_ph_transition_CR_Single} Called directly from R, Defines the control variables and calls the regression function
}
