% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cox_Refresh_R_SIDES}
\alias{Cox_Refresh_R_SIDES}
\title{Utility function to refresh side matrices for Cox Omnibus
\code{Cox_Refresh_R_SIDES} Called to fresh repeated sum calculation matrices}
\usage{
Cox_Refresh_R_SIDES(
  reqrdnum,
  ntime,
  Rls1,
  Rls2,
  Rls3,
  Lls1,
  Lls2,
  Lls3,
  STRATA_vals,
  strata_bool,
  single_bool
)
}
\arguments{
\item{reqrdnum}{total number of free parameters}

\item{ntime}{number of risk groups}

\item{Rls1}{First Risk sum storage}

\item{Rls2}{First Risk sum derivative storage}

\item{Rls3}{First Risk sum second derivative storage}

\item{Lls1}{Second Risk sum storage}

\item{Lls2}{Second Risk sum derivative storage}

\item{Lls3}{Second Risk sum second derivative storage}

\item{STRATA_vals}{vector of strata identifier values}

\item{strata_bool}{boolean for stratification}

\item{single_bool}{boolean for single calculation without derivatives}
}
\value{
Updates matrices in place: risk storage matrices
}
\description{
Utility function to refresh side matrices for Cox Omnibus
\code{Cox_Refresh_R_SIDES} Called to fresh repeated sum calculation matrices
}
