% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cox_Regression.R
\name{Cox_Relative_Risk}
\alias{Cox_Relative_Risk}
\title{Calculates hazard ratios for a reference vector}
\usage{
Cox_Relative_Risk(
  df,
  time1 = "\%trunc\%",
  time2 = "\%trunc\%",
  event0 = "event",
  names = c("CONST"),
  term_n = c(0),
  tform = "loglin",
  keep_constant = c(0),
  a_n = c(0),
  modelform = "M",
  control = list(),
  model_control = list()
)
}
\arguments{
\item{df}{a data.table containing the columns of interest}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{event0}{column used for event status}

\item{names}{columns for elements of the model, used to identify data columns}

\item{term_n}{term numbers for each element of the model}

\item{tform}{list of string function identifiers, used for linear/step}

\item{keep_constant}{binary values to denote which parameters to change}

\item{a_n}{list of initial parameter values, used to determine the number of parameters. May be either a list of vectors or a single vector.}

\item{modelform}{string specifying the model type: M, ME, A, PA, PAE, GMIX, GMIX-R, GMIX-E}

\item{control}{list of parameters controlling the convergence, see the Control_Options vignette for details}

\item{model_control}{controls which alternative model options are used, see the Control_Options vignette for further details}
}
\value{
returns a list of the final results
}
\description{
\code{RunCoxRegression} uses user provided data,  vectors specifying the model,
and options to calculate relative risk for every row in the provided data
}
\seealso{
Other Plotting Wrapper Functions: 
\code{\link{RunCoxPlots}()},
\code{\link{plot.coxres}()}
}
\concept{Plotting Wrapper Functions}
