% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTypes.r
\name{CoxKaplanMeier}
\alias{CoxKaplanMeier}
\title{calculates and plots Kaplan-Mieier survival plots
\code{CoxKaplanMeier} uses user provided data, columns, and identifier to create plots, plots the survival for subjects above/below the mean column value for each column}
\usage{
CoxKaplanMeier(
  verbose,
  verbosec,
  studyID,
  names,
  df,
  event,
  time1,
  time2,
  tu,
  Term_n,
  tform,
  a_n,
  er,
  fir,
  der_iden,
  modelform,
  control,
  keep_constant,
  Plot_Type,
  age_unit
)
}
\arguments{
\item{verbose}{boolean identifying if extra plotting information should be written to the console}

\item{verbosec}{boolean identifying if extra regression information should be written to the console}

\item{studyID}{column identifying each individual}

\item{names}{columns names for elements of the model used to identify data columns}

\item{df}{data used for regression}

\item{event}{column used for event status}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{tu}{unique event times}

\item{Term_n}{term numbers for each element of the model}

\item{tform}{subterm type for each element of the model}

\item{a_n}{starting parameters for regression}

\item{er}{standard deviation for the parameters}

\item{fir}{term number for the intial term used for models of the form T0*f(Ti) in which the order matters}

\item{der_iden}{number for the subterm to test derivative at only used for testing runs with a single varying parameter}

\item{modelform}{string specifying the model type}

\item{control}{list of parameters controlling the convergence}

\item{keep_constant}{vector of 0/1 to identify parameters to force to be constant}

\item{Plot_Type}{list of parameters controlling the plot options}

\item{age_unit}{age unit}
}
\value{
saves the plots in the current directory and returns a string that it passed
}
\description{
calculates and plots Kaplan-Mieier survival plots
\code{CoxKaplanMeier} uses user provided data, columns, and identifier to create plots, plots the survival for subjects above/below the mean column value for each column
}
