% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTypes.r
\name{CoxStratifiedSurvival}
\alias{CoxStratifiedSurvival}
\title{calculates and plots survival curves for each unique value of the stratification column
\code{CoxStratifiedSurvival} uses user provided data, columns, and identifier to calculate the survival fraction for each strata}
\usage{
CoxStratifiedSurvival(
  verbose,
  df,
  event,
  time1,
  time2,
  names,
  Term_n,
  tform,
  a_n,
  er,
  fir,
  der_iden,
  modelform,
  control,
  keep_constant,
  Plot_Type,
  Strat_Col,
  time_lims,
  age_unit
)
}
\arguments{
\item{verbose}{boolean to control if additional information is printed to the console}

\item{df}{data used for regression}

\item{event}{column used for event status}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{names}{columns names for elements of the model used to identify data columns}

\item{Term_n}{term numbers for each element of the model}

\item{tform}{subterm type for each element of the model}

\item{a_n}{starting parameters for regression}

\item{er}{standard deviation for the parameters}

\item{fir}{term number for the intial term used for models of the form T0*f(Ti) in which the order matters}

\item{der_iden}{number for the subterm to test derivative at only used for testing runs with a single varying parameter}

\item{modelform}{string specifying the model type}

\item{control}{list of parameters controlling the convergence}

\item{keep_constant}{vector of 0/1 to identify parameters to force to be constant}

\item{Plot_Type}{list of parameters controlling the plot options}

\item{Strat_Col}{column to stratify by}

\item{time_lims}{limits for x axis of plot}

\item{age_unit}{age unit}
}
\value{
saves the plots in the current directory and returns a string that it passed
}
\description{
calculates and plots survival curves for each unique value of the stratification column
\code{CoxStratifiedSurvival} uses user provided data, columns, and identifier to calculate the survival fraction for each strata
}
