% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTypes.r
\name{CoxSurvival}
\alias{CoxSurvival}
\title{calculates and plots survival plots of the estimated baseline
\code{CoxSurvival} uses user provided data, columns, and identifier to create plots}
\usage{
CoxSurvival(t, h, ch, surv, Plot_Name, verbose, time_lims)
}
\arguments{
\item{t}{event times}

\item{h}{hazards of baseline}

\item{ch}{cumulative hazards of baseline}

\item{surv}{survival fraction of baseline}

\item{Plot_Name}{plot identifier}

\item{verbose}{boolean controlling additional printing}

\item{time_lims}{limits for x axis of plot}
}
\value{
saves the plots in the current directory and returns a string that it passed
}
\description{
calculates and plots survival plots of the estimated baseline
\code{CoxSurvival} uses user provided data, columns, and identifier to create plots
}
