% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Def_Control}
\alias{Def_Control}
\title{Automatically assigns missing control values
\code{Def_Control} checks and assigns default values}
\usage{
Def_Control(control)
}
\arguments{
\item{control}{list of control parameters}
}
\value{
returns a filled list
}
\description{
Automatically assigns missing control values
\code{Def_Control} checks and assigns default values
}
