% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Def_model_control}
\alias{Def_model_control}
\title{Automatically assigns missing model control values}
\usage{
Def_model_control(control)
}
\arguments{
\item{control}{list of parameters controlling the convergence, see Def_Control() for options or vignette("Control_Options")}
}
\value{
returns a filled list
}
\description{
\code{Def_model_control} checks and assigns default values
}
\examples{
library(data.table)
control <- list("single"=TRUE)
control <- Def_model_control(control)

}
\seealso{
Other Data Cleaning Functions: 
\code{\link{Check_Dupe_Columns}()},
\code{\link{Check_Trunc}()},
\code{\link{Check_Verbose}()},
\code{\link{Correct_Formula_Order}()},
\code{\link{Date_Shift}()},
\code{\link{Def_Control}()},
\code{\link{Def_Control_Guess}()},
\code{\link{Def_modelform_fix}()},
\code{\link{Joint_Multiple_Events}()},
\code{\link{Replace_Missing}()},
\code{\link{Time_Since}()},
\code{\link{factorize}()},
\code{\link{factorize_par}()},
\code{\link{gen_time_dep}()},
\code{\link{interact_them}()}
}
\concept{Data Cleaning Functions}
