% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{Event_Count_Gen}
\alias{Event_Count_Gen}
\title{uses a table, list of categories, and list of event summaries to generate person-count tables}
\usage{
Event_Count_Gen(table, categ = list(), events = list())
}
\arguments{
\item{table}{dataframe with every category/event column needed}

\item{categ}{list with category columns and methods, methods can be either strings or lists of boundaries}

\item{events}{list of columns to summarize, supports counts and means and renaming the summary column}
}
\value{
returns a grouped table and a list of category boundaries used
}
\description{
\code{Event_Count_Gen} generates event-count tables
}
\examples{
library(data.table)
a <- c(0, 1, 2, 3, 4, 5, 6)
b <- c(1, 2, 3, 4, 5, 6, 7)
c <- c(0, 1, 0, 0, 0, 1, 0)
table <- data.table(
  a = a,
  b = b,
  c = c
)
categ <- list(
  a = "0/3/5]7",
  b = list(
    lower = c(-1, 3, 6),
    upper = c(3, 6, 10),
    name = c("low", "medium", "high")
  )
)
event <- list(
  c = "count AS cases",
  a = "mean", b = "mean"
)
e <- Event_Count_Gen(table, categ, event)
}
\seealso{
Other Table Generation Functions:
\code{\link[=Event_Time_Gen]{Event_Time_Gen()}}
}
\concept{Table Generation Functions}
