% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{factorize_par}
\alias{factorize_par}
\title{Splits a parameter into factors in parallel
\code{factorize_par} uses user provided list of columns to define new parameter for each unique value and update the data.table.
Not for interaction terms}
\usage{
factorize_par(
  df,
  col_list,
  verbose = FALSE,
  nthreads = as.numeric(detectCores())
)
}
\arguments{
\item{df}{a data.table containing the columns of interest}

\item{col_list}{an array of column names that should have factor terms defined}

\item{verbose}{verbosity argument, for error returns}

\item{nthreads}{number of threads to use}
}
\value{
returns a list with two named fields. df for the updated dataframe, and cols for the new column names
}
\description{
Splits a parameter into factors in parallel
\code{factorize_par} uses user provided list of columns to define new parameter for each unique value and update the data.table.
Not for interaction terms
}
