% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{gcc_version}
\alias{gcc_version}
\title{Checks default c++ compiler}
\usage{
gcc_version()
}
\value{
returns a string representation of gcc, clang, or c++ output
}
\description{
\code{gcc_version} Checks default c++ compiler, part of configuration script
}
