% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gen_Fac_Par}
\alias{Gen_Fac_Par}
\title{Generates factored columns in parallel
\code{Gen_Fac_Par} Called directly from R, returns a matrix with factored columns}
\usage{
Gen_Fac_Par(df0, vals, cols, nthreads)
}
\arguments{
\item{df0}{matrix of covariate values}

\item{vals}{list of values for each column, single continuous list}

\item{cols}{list of column identifiers, single continuous list}

\item{nthreads}{number of threads available}
}
\value{
saves a dataframe to be used with time-dependent covariate analysis
}
\description{
Generates factored columns in parallel
\code{Gen_Fac_Par} Called directly from R, returns a matrix with factored columns
}
