% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gen_Strat_Weight}
\alias{Gen_Strat_Weight}
\title{Generates weightings for stratified poisson regression
\code{Gen_Strat_Weight} Called from within c++, assigns vector of weights}
\usage{
Gen_Strat_Weight(
  modelform,
  dfs,
  PyrC,
  s_weights,
  nthreads,
  tform,
  Term_n,
  term_tot
)
}
\arguments{
\item{modelform}{string model identifier}

\item{dfs}{Matrix with stratification columns, assumed to be binary and mutually exclusive}

\item{PyrC}{matrix of person-years and event counts}

\item{s_weights}{vector of weights for every row}

\item{nthreads}{number of threads available}

\item{tform}{subterm types}

\item{Term_n}{Term numbers}

\item{term_tot}{total number of terms}
}
\value{
assigns weight in place and returns nothing
}
\description{
Generates weightings for stratified poisson regression
\code{Gen_Strat_Weight} Called from within c++, assigns vector of weights
}
