% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{get_conf_int}
\alias{get_conf_int}
\title{Calculates a chi-squared value, used in a non-functioning bounds formula
\code{get_conf_int} uses an interval parameter to return the score statistic}
\usage{
get_conf_int(alpha = 0.95)
}
\arguments{
\item{alpha}{decimal 1 - (confidence interval)}
}
\value{
returns the score statistic
}
\description{
Calculates a chi-squared value, used in a non-functioning bounds formula
\code{get_conf_int} uses an interval parameter to return the score statistic
}
