% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{get_form_list}
\alias{get_form_list}
\title{Interprets a Colossus formula and makes necessary changes to data, returns list not class}
\usage{
get_form_list(surv_obj, model_obj, df)
}
\arguments{
\item{surv_obj}{output from get_form_surv, list of survival values}

\item{model_obj}{output from get_form_risk, list of risk factor model values}

\item{df}{a data.table containing the columns of interest}
}
\value{
returns the list of model values
}
\description{
\code{get_form_list} uses a formula and data.table, to fully describe the model
for a Colossus regression function and returns a list
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCaseConStrataSurv}()},
\code{\link{ColossusCaseConSurv}()},
\code{\link{ColossusCaseConTimeStrataSurv}()},
\code{\link{ColossusCaseConTimeSurv}()},
\code{\link{ColossusCoxStrataSurv}()},
\code{\link{ColossusCoxSurv}()},
\code{\link{ColossusExpressionCall}()},
\code{\link{ColossusFineGrayStrataSurv}()},
\code{\link{ColossusFineGraySurv}()},
\code{\link{ColossusLogitSurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()},
\code{\link{get_form_joint}()},
\code{\link{get_form_risk}()},
\code{\link{get_form_surv}()}
}
\concept{Formula Interpretation}
