% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{get_form_surv}
\alias{get_form_surv}
\title{Interprets the survival side of a formula}
\usage{
get_form_surv(surv_obj, df)
}
\arguments{
\item{surv_obj}{output from get_form_surv, list of survival values}

\item{df}{a data.table containing the columns of interest}
}
\value{
returns the left hand side of the formula reading
}
\description{
\code{get_form_surv} interprets the LHS of a formula
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCaseConStrataSurv}()},
\code{\link{ColossusCaseConSurv}()},
\code{\link{ColossusCaseConTimeStrataSurv}()},
\code{\link{ColossusCaseConTimeSurv}()},
\code{\link{ColossusCoxStrataSurv}()},
\code{\link{ColossusCoxSurv}()},
\code{\link{ColossusExpressionCall}()},
\code{\link{ColossusFineGrayStrataSurv}()},
\code{\link{ColossusFineGraySurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()},
\code{\link{get_form_joint}()},
\code{\link{get_form_list}()},
\code{\link{get_form_risk}()}
}
\concept{Formula Interpretation}
