% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{interact_them}
\alias{interact_them}
\title{Defines Interactions
\code{interact_them} uses user provided interactions define interaction terms and update the data.table. assumes interaction is "+" or "*" and applies basic anti-aliasing to avoid duplicates}
\usage{
interact_them(df, interactions, new_names, verbose = FALSE)
}
\arguments{
\item{df}{a datatable containing the columns of interest}

\item{interactions}{array of strings, each one is of form term1?*?term2" for term1 interaction of type * with term2, "?" dlimits}

\item{new_names}{list of interaction names to use instead of default, default used if entry is ''}

\item{verbose}{verbosity}
}
\value{
returns a list with two named fields. df for the updated dataframe, and cols for the new column names
}
\description{
Defines Interactions
\code{interact_them} uses user provided interactions define interaction terms and update the data.table. assumes interaction is "+" or "*" and applies basic anti-aliasing to avoid duplicates
}
