% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Intercept_Bound}
\alias{Intercept_Bound}
\title{Utility function to keep intercept parameters within the range of possible values
\code{Intercept_Bound} Called to update the parameter list in the event that intercepts leave the bounds of possible values}
\usage{
Intercept_Bound(
  nthreads,
  totalnum,
  beta_0,
  dbeta,
  dfc,
  df0,
  KeepConstant,
  debugging,
  tform
)
}
\arguments{
\item{nthreads}{number of threads}

\item{totalnum}{total number of parameter}

\item{beta_0}{parameter list}

\item{dbeta}{parameter change vector}

\item{dfc}{covariate column numbers}

\item{df0}{covariate matrix}

\item{KeepConstant}{vector of parameters to keep constant}

\item{debugging}{debugging boolean}

\item{tform}{subterm type}
}
\value{
Updates vector in place: parameter vector
}
\description{
Utility function to keep intercept parameters within the range of possible values
\code{Intercept_Bound} Called to update the parameter list in the event that intercepts leave the bounds of possible values
}
