% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Interpret_Output}
\alias{Interpret_Output}
\title{Prints a regression output clearly}
\usage{
Interpret_Output(out_list, digits = 2)
}
\arguments{
\item{out_list}{list output from a regression, used to build results table and pull out convergence values}

\item{digits}{digits used for printing results}
}
\value{
return nothing, prints the results to console
}
\description{
\code{Interpret_Output} uses the list output from a regression, prints off a table of results and summarizes the score and convergence.
}
