% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTypes.r
\name{K_Smooth}
\alias{K_Smooth}
\title{calculates a smoothing factor
\code{K_Smooth} calculates 3/4 of (1-x)^2}
\usage{
K_Smooth(x)
}
\arguments{
\item{x}{point to smooth using}
}
\value{
returns the smoothing factor
}
\description{
calculates a smoothing factor
\code{K_Smooth} calculates 3/4 of (1-x)^2
}
