% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_run.R
\name{LikelihoodBound}
\alias{LikelihoodBound}
\title{Generic likelihood boundary calculation function}
\usage{
LikelihoodBound(x, df, curve_control = list(), control = list(), ...)
}
\arguments{
\item{x}{result object from a regression, class coxres or poisres}

\item{df}{a data.table containing the columns of interest}

\item{curve_control}{a list of control options for the likelihood boundary regression. See the \href{../articles/Control_Options.html}{control options vignette} for details.}

\item{control}{list of parameters controlling the convergence, see the \href{../articles/Control_Options.html}{control options vignette} for details}

\item{...}{extended for other necessary parameters}
}
\value{
returns a summary of the likelihood boundary
}
\description{
\code{LikelihoodBound} Generic likelihood boundary calculation function. Used
by the coxres, poisres, and logitres objects to calculate likelihood based
confidence intervals.
}
\seealso{
Other Likelihood Boundaries:
\code{\link[=LikelihoodBound.coxres]{LikelihoodBound.coxres()}},
\code{\link[=LikelihoodBound.logitres]{LikelihoodBound.logitres()}},
\code{\link[=LikelihoodBound.poisres]{LikelihoodBound.poisres()}}
}
\concept{Likelihood Boundaries}
