% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{Linked_Dose_Formula}
\alias{Linked_Dose_Formula}
\title{Calculates Full Parameter list for Special Dose Formula}
\usage{
Linked_Dose_Formula(tforms, paras)
}
\arguments{
\item{tforms}{list of formula types}

\item{paras}{list of formula parameters}
}
\value{
returns list of full parameters
}
\description{
\code{Linked_Dose_Formula} Calculates all parameters for linear-quadratic and linear-exponential linked formulas
}
\examples{
library(data.table)
tforms <- list(cov_0 = "quad", cov_1 = "exp")
paras <- list(cov_0 = c(1, 3.45), cov_1 = c(1.2, 4.5, 0.1))
full_paras <- Linked_Dose_Formula(tforms, paras)
}
\seealso{
Other Piece-wise Assistance Functions:
\code{\link[=Linked_Lin_Exp_Para]{Linked_Lin_Exp_Para()}}
}
\concept{Piece-wise Assistance Functions}
