% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{Linked_Lin_Exp_Para}
\alias{Linked_Lin_Exp_Para}
\title{Calculates The Additional Parameter For a linear-exponential formula with known maximum}
\usage{
Linked_Lin_Exp_Para(y, a0, a1_goal)
}
\arguments{
\item{y}{point formula switch}

\item{a0}{linear slope}

\item{a1_goal}{exponential maximum desired}
}
\value{
returns parameter used by Colossus
}
\description{
\code{Linked_Lin_Exp_Para} Calculates what the additional parameter would be for a desired maximum
}
\examples{
library(data.table)
y <- 7.6
a0 <- 1.2
a1_goal <- 15
full_paras <- Linked_Lin_Exp_Para(y, a0, a1_goal)
}
\seealso{
Other Piece-wise Assistance Functions:
\code{\link[=Linked_Dose_Formula]{Linked_Dose_Formula()}}
}
\concept{Piece-wise Assistance Functions}
