% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLik_Cox_PH}
\alias{LogLik_Cox_PH}
\title{Primary Cox PH regression
\code{LogLik_Cox_PH} Performs the calls to calculation functions, Structures the Cox PH regression, With verbose option prints out time stamps and intermediate sums of terms and derivatives}
\usage{
LogLik_Cox_PH(
  Term_n,
  tform,
  a_n,
  x_all,
  dfc,
  fir,
  der_iden,
  modelform,
  lr,
  maxiter,
  halfmax,
  epsilon,
  dbeta_cap,
  abs_max,
  dose_abs_max,
  deriv_epsilon,
  df_groups,
  tu,
  double_step,
  change_all,
  verbose,
  debugging,
  KeepConstant,
  term_tot,
  ties_method,
  nthreads
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{x_all}{covariate matrix}

\item{dfc}{covariate column numbers}

\item{fir}{first term number}

\item{der_iden}{subterm number for derivative tests}

\item{modelform}{model string}

\item{lr}{learning rate for newton step toward 0 derivative}

\item{maxiter}{maximum number of iterations}

\item{halfmax}{maximum number of half steps}

\item{epsilon}{minimum acceptable maximum parameter change}

\item{dbeta_cap}{learning rate for newton step toward 0 log-likelihood}

\item{abs_max}{Maximum allowed parameter change}

\item{dose_abs_max}{Maximum allowed threshold parameter change}

\item{deriv_epsilon}{threshold for near-zero derivative}

\item{df_groups}{matrix with time and event information}

\item{tu}{event times}

\item{double_step}{controls the step calculation, 0 for independent changes, 1 for solving b=Ax with complete matrices}

\item{change_all}{boolean if every parameter is being updated}

\item{verbose}{verbosity boolean}

\item{debugging}{debugging boolean}

\item{KeepConstant}{vector identifying constant parameters}

\item{term_tot}{total number of terms}

\item{ties_method}{ties method}

\item{nthreads}{number of threads to use}
}
\value{
List of results: Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, model information
}
\description{
Primary Cox PH regression
\code{LogLik_Cox_PH} Performs the calls to calculation functions, Structures the Cox PH regression, With verbose option prints out time stamps and intermediate sums of terms and derivatives
}
