% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLik_Cox_PH_null}
\alias{LogLik_Cox_PH_null}
\title{Primary Cox PH null model function
\code{LogLik_Cox_PH_null} Performs the calls to calculation functions, Structures the Cox PH null model, With verbose option prints out time stamps and intermediate sums of terms and derivatives}
\usage{
LogLik_Cox_PH_null(df_groups, tu, verbose, ties_method)
}
\arguments{
\item{df_groups}{status/time matrix}

\item{tu}{vector of event times}

\item{verbose}{verbose boolean}

\item{ties_method}{tied event method}
}
\value{
List of results: Log-likelihood of optimum, AIC
}
\description{
Primary Cox PH null model function
\code{LogLik_Cox_PH_null} Performs the calls to calculation functions, Structures the Cox PH null model, With verbose option prints out time stamps and intermediate sums of terms and derivatives
}
