% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLik_Cox_PH_Single}
\alias{LogLik_Cox_PH_Single}
\title{Primary Cox PH calculation
\code{LogLik_Cox_PH_Single} Calculates just the log-likelihood}
\usage{
LogLik_Cox_PH_Single(
  Term_n,
  tform,
  a_n,
  x_all,
  dfc,
  fir,
  modelform,
  df_groups,
  tu,
  verbose,
  debugging,
  KeepConstant,
  term_tot,
  ties_method,
  nthreads
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{x_all}{covariate matrix}

\item{dfc}{covariate column numbers}

\item{fir}{first term number}

\item{modelform}{model string}

\item{df_groups}{matrix with time and event information}

\item{tu}{event times}

\item{verbose}{verbosity boolean}

\item{debugging}{debugging boolean}

\item{KeepConstant}{vector identifying constant parameters}

\item{term_tot}{total number of terms}

\item{ties_method}{ties method}

\item{nthreads}{number of threads to use}
}
\value{
List of results: Log-likelihood, parameter list, AIC, model information
}
\description{
Primary Cox PH calculation
\code{LogLik_Cox_PH_Single} Calculates just the log-likelihood
}
