% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLik_Cox_PH_STRATA_SINGLE_CR}
\alias{LogLik_Cox_PH_STRATA_SINGLE_CR}
\title{Primary Cox PH regression with STRATA effect, calculating only the Log-Likelihood, and competing risks event=2
\code{LogLik_Cox_PH_STRATA_SINGLE_CR} Performs the calls to calculation functions, Structures the Cox PH regression, With verbose option prints out time stamps and intermediate sums of terms and derivatives}
\usage{
LogLik_Cox_PH_STRATA_SINGLE_CR(
  Term_n,
  tform,
  a_n,
  x_all,
  dfc,
  fir,
  modelform,
  df_groups,
  tu,
  cens_weight,
  cens_thres,
  verbose,
  debugging,
  KeepConstant,
  term_tot,
  ties_method,
  STRATA_vals,
  nthreads
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{x_all}{covariate matrix}

\item{dfc}{covariate column numbers}

\item{fir}{first term number}

\item{modelform}{model string}

\item{df_groups}{matrix with time and event information}

\item{tu}{event times}

\item{cens_weight}{censoring weight list}

\item{cens_thres}{threshold to add competing event to risk group}

\item{verbose}{verbosity boolean}

\item{debugging}{debugging boolean}

\item{KeepConstant}{vector identifying constant parameters}

\item{term_tot}{total number of terms}

\item{ties_method}{ties method}

\item{STRATA_vals}{vector of strata identifier values}

\item{nthreads}{number of threads to use}
}
\value{
List of results: Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, model information
}
\description{
Primary Cox PH regression with STRATA effect, calculating only the Log-Likelihood, and competing risks event=2
\code{LogLik_Cox_PH_STRATA_SINGLE_CR} Performs the calls to calculation functions, Structures the Cox PH regression, With verbose option prints out time stamps and intermediate sums of terms and derivatives
}
