% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLik_Pois_Omnibus}
\alias{LogLik_Pois_Omnibus}
\title{Primary poisson regression with multiple starting points and optional combinations of stratification and no derivative calculation.
\code{LogLik_Pois_Omnibus} Performs the calls to calculation functions, Structures the poisson regression, With verbose option prints out time stamps and intermediate sums of terms and derivatives}
\usage{
LogLik_Pois_Omnibus(
  PyrC,
  Term_n,
  tform,
  a_ns,
  x_all,
  dfc,
  fir,
  der_iden,
  modelform,
  lr,
  maxiters,
  guesses,
  halfmax,
  epsilon,
  dbeta_cap,
  abs_max,
  dose_abs_max,
  deriv_epsilon,
  double_step,
  change_all,
  verbose,
  debugging,
  KeepConstant,
  term_tot,
  nthreads,
  dfs,
  strata_bool,
  single_bool
)
}
\arguments{
\item{PyrC}{person-year matrix}

\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_ns}{matrix starting values}

\item{x_all}{covariate matrix}

\item{dfc}{covariate column numbers}

\item{fir}{first term number}

\item{der_iden}{subterm number for derivative tests}

\item{modelform}{model string}

\item{lr}{learning rate for newton step toward 0 derivative}

\item{maxiters}{list of maximum number of iterations}

\item{guesses}{number of initial guesses}

\item{halfmax}{maximum number of half steps}

\item{epsilon}{minimum acceptable maximum parameter change}

\item{dbeta_cap}{learning rate for newton step toward 0 log-likelihood}

\item{abs_max}{Maximum allowed parameter change}

\item{dose_abs_max}{Maximum allowed threshold parameter change}

\item{deriv_epsilon}{threshold for near-zero derivative}

\item{double_step}{controls the step calculation, 0 for independent changes, 1 for solving b=Ax with complete matrices}

\item{change_all}{boolean if every parameter is being updated}

\item{verbose}{verbosity boolean}

\item{debugging}{debugging boolean}

\item{KeepConstant}{vector identifying constant parameters}

\item{term_tot}{total number of terms}

\item{nthreads}{number of threads to use}

\item{dfs}{matrix of stratification variables}

\item{strata_bool}{boolean for stratification}

\item{single_bool}{boolean for single calculation without derivatives}
}
\value{
List of final results: Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, model information
}
\description{
Primary poisson regression with multiple starting points and optional combinations of stratification and no derivative calculation.
\code{LogLik_Pois_Omnibus} Performs the calls to calculation functions, Structures the poisson regression, With verbose option prints out time stamps and intermediate sums of terms and derivatives
}
