% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLik_Poisson}
\alias{LogLik_Poisson}
\title{Primary poisson regression function
\code{LogLik_Poisson} Performs the calls to calculation functions, Structures the poisson regression, With verbose option prints out time stamps and intermediate sums of terms and derivatives}
\usage{
LogLik_Poisson(
  PyrC,
  Term_n,
  tform,
  a_n,
  x_all,
  dfc,
  fir,
  der_iden,
  modelform,
  lr,
  maxiter,
  halfmax,
  epsilon,
  dbeta_cap,
  abs_max,
  dose_abs_max,
  deriv_epsilon,
  change_all,
  verbose,
  debugging,
  KeepConstant,
  term_tot
)
}
\arguments{
\item{PyrC}{person-year matrix}

\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{x_all}{covariate matrix}

\item{dfc}{covariate column numbers}

\item{fir}{first term number}

\item{der_iden}{subterm number for derivative tests}

\item{modelform}{model string}

\item{lr}{learning rate for newton step toward 0 derivative}

\item{maxiter}{maximum number of iterations}

\item{halfmax}{maximum number of half steps}

\item{epsilon}{minimum acceptable maximum parameter change}

\item{dbeta_cap}{learning rate for newton step toward 0 log-likelihood}

\item{abs_max}{Maximum allowed parameter change}

\item{dose_abs_max}{Maximum allowed threshold parameter change}

\item{deriv_epsilon}{threshold for near-zero derivative}

\item{change_all}{boolean if every parameter is being updated}

\item{verbose}{verbosity boolean}

\item{debugging}{debugging boolean}

\item{KeepConstant}{vector identifying constant parameters}

\item{term_tot}{total number of terms}
}
\value{
List of results: Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, deviance, model information
}
\description{
Primary poisson regression function
\code{LogLik_Poisson} Performs the calls to calculation functions, Structures the poisson regression, With verbose option prints out time stamps and intermediate sums of terms and derivatives
}
