% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLik_Poisson_Single}
\alias{LogLik_Poisson_Single}
\title{Primary poisson calculation function
\code{LogLik_Poisson_Single} Performs the calls to calculation functions, Structures the poisson calculation}
\usage{
LogLik_Poisson_Single(
  PyrC,
  Term_n,
  tform,
  a_n,
  x_all,
  dfc,
  fir,
  modelform,
  verbose,
  debugging,
  KeepConstant,
  term_tot,
  nthreads
)
}
\arguments{
\item{PyrC}{person-year matrix}

\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{x_all}{covariate matrix}

\item{dfc}{covariate column numbers}

\item{fir}{first term number}

\item{modelform}{model string}

\item{verbose}{verbosity boolean}

\item{debugging}{debugging boolean}

\item{KeepConstant}{vector identifying constant parameters}

\item{term_tot}{total number of terms}

\item{nthreads}{number of threads to use}
}
\value{
List of results: Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, deviance, model information
}
\description{
Primary poisson calculation function
\code{LogLik_Poisson_Single} Performs the calls to calculation functions, Structures the poisson calculation
}
