% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Make_Groups_STRATA_CR}
\alias{Make_Groups_STRATA_CR}
\title{Utility function to define risk groups with STRATA and competing risks
\code{Make_Groups_STRATA_CR} Called to update lists of risk groups, Uses list of event times and row time/event information, Matrices store starting/stopping row indices for each group , adds competing risks}
\usage{
Make_Groups_STRATA_CR(
  ntime,
  df_m,
  RiskFail,
  RiskGroup,
  tu,
  nthreads,
  debugging,
  STRATA_vals,
  cens_weight,
  cens_cutoff
)
}
\arguments{
\item{ntime}{number of event times}

\item{df_m}{event/time matrix}

\item{RiskFail}{Matrix of event rows for each event time}

\item{RiskGroup}{vectors of strings with rows at risk for each event time}

\item{tu}{Event time vector}

\item{nthreads}{number of threads}

\item{debugging}{debugging boolean}

\item{STRATA_vals}{vector of strata identifier values}

\item{cens_weight}{vector with censoring weights}

\item{cens_cutoff}{double threshold for adding competing risk to risk group, not implemented}
}
\value{
Updates matrices in place: Matrix of event rows for each event time, vectors of strings with rows at risk for each event time
}
\description{
Utility function to define risk groups with STRATA and competing risks
\code{Make_Groups_STRATA_CR} Called to update lists of risk groups, Uses list of event times and row time/event information, Matrices store starting/stopping row indices for each group , adds competing risks
}
