% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Make_Risks}
\alias{Make_Risks}
\title{Utility function to calculate the risk and risk ratios
\code{Make_Risks} Called to update risk matrices, Splits into cases based on model form, Uses lists of term numbers and types to apply different derivative formulas}
\usage{
Make_Risks(
  modelform,
  tform,
  Term_n,
  totalnum,
  fir,
  T0,
  Td0,
  Tdd0,
  Te,
  R,
  Rd,
  Rdd,
  Dose,
  nonDose,
  TTerm,
  nonDose_LIN,
  nonDose_PLIN,
  nonDose_LOGLIN,
  RdR,
  RddR,
  nthreads,
  debugging
)
}
\arguments{
\item{modelform}{Model string}

\item{tform}{subterm types}

\item{Term_n}{term numbers}

\item{totalnum}{total number of terms}

\item{fir}{first term number}

\item{T0}{Term by subterm matrix}

\item{Td0}{Term by subterm derivative matrix}

\item{Tdd0}{Term by subterm second derivative matrix}

\item{Te}{Temporary term storage matrix}

\item{R}{Risk matrix}

\item{Rd}{Risk first derivative matrix}

\item{Rdd}{Risk second derivative matrix}

\item{Dose}{Dose term matrix}

\item{nonDose}{nonDose term matrix}

\item{TTerm}{Total term matrix}

\item{nonDose_LIN}{Linear term matrix}

\item{nonDose_PLIN}{Product linear term matrix}

\item{nonDose_LOGLIN}{Loglinear term matrix}

\item{RdR}{Risk to first derivative ratio matrix}

\item{RddR}{Risk to second derivative ratio matrix}

\item{nthreads}{number of threads to use}

\item{debugging}{debugging boolean}
}
\value{
Updates matrices in place: Risk, Risk ratios
}
\description{
Utility function to calculate the risk and risk ratios
\code{Make_Risks} Called to update risk matrices, Splits into cases based on model form, Uses lists of term numbers and types to apply different derivative formulas
}
