% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Make_Risks_Basic}
\alias{Make_Risks_Basic}
\title{Utility function to calculate the risk and risk ratios for the basic model
\code{Make_Risks_Basic} Called to update risk matrices, Splits into cases based on model form, Uses lists of term numbers and types to apply different derivative formulas}
\usage{
Make_Risks_Basic(
  totalnum,
  T0,
  R,
  Rd,
  Rdd,
  RdR,
  nthreads,
  debugging,
  df0,
  dfc,
  KeepConstant
)
}
\arguments{
\item{totalnum}{total number of terms}

\item{T0}{Term by subterm matrix}

\item{R}{Risk matrix}

\item{Rd}{Risk first derivative matrix}

\item{Rdd}{Risk second derivative matrix}

\item{RdR}{Risk to first derivative ratio matrix}

\item{nthreads}{number of threads to use}

\item{debugging}{debugging boolean}

\item{df0}{covariate matrix}

\item{dfc}{covariate column numbers}

\item{KeepConstant}{vector identifying constant parameters}
}
\value{
Updates matrices in place: Risk, Risk ratios
}
\description{
Utility function to calculate the risk and risk ratios for the basic model
\code{Make_Risks_Basic} Called to update risk matrices, Splits into cases based on model form, Uses lists of term numbers and types to apply different derivative formulas
}
