% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Make_Risks_Single}
\alias{Make_Risks_Single}
\title{Utility function to calculate the risk, but not derivatives
\code{Make_Risks_Single} Called to update risk matrices, Splits into cases based on model form}
\usage{
Make_Risks_Single(
  modelform,
  tform,
  Term_n,
  totalnum,
  fir,
  T0,
  Te,
  R,
  Dose,
  nonDose,
  TTerm,
  nonDose_LIN,
  nonDose_PLIN,
  nonDose_LOGLIN,
  nthreads,
  debugging
)
}
\arguments{
\item{modelform}{Model string}

\item{tform}{subterm types}

\item{Term_n}{term numbers}

\item{totalnum}{total number of terms}

\item{fir}{first term number}

\item{T0}{Term by subterm matrix}

\item{Te}{Temporary term storage matrix}

\item{R}{Risk matrix}

\item{Dose}{Dose term matrix}

\item{nonDose}{nonDose term matrix}

\item{TTerm}{Total term matrix}

\item{nonDose_LIN}{Linear term matrix}

\item{nonDose_PLIN}{Product linear term matrix}

\item{nonDose_LOGLIN}{Loglinear term matrix}

\item{nthreads}{number of threads to use}

\item{debugging}{debugging boolean}
}
\value{
Updates matrices in place: Risk, Risk ratios
}
\description{
Utility function to calculate the risk, but not derivatives
\code{Make_Risks_Single} Called to update risk matrices, Splits into cases based on model form
}
