% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Make_Risks_Weighted}
\alias{Make_Risks_Weighted}
\title{Utility function to calculate the risk and risk ratios with a weighting applied
\code{Make_Risks_Weighted} Called to update weighted risk matrices, Splits into cases based on model form, Uses lists of term numbers and types to apply different derivative formulas}
\usage{
Make_Risks_Weighted(
  modelform,
  tform,
  Term_n,
  totalnum,
  fir,
  s_weights,
  T0,
  Td0,
  Tdd0,
  Te,
  R,
  Rd,
  Rdd,
  Dose,
  nonDose,
  TTerm,
  nonDose_LIN,
  nonDose_PLIN,
  nonDose_LOGLIN,
  RdR,
  RddR,
  nthreads,
  debugging,
  KeepConstant,
  gmix_theta,
  gmix_term
)
}
\arguments{
\item{modelform}{string model identifier}

\item{tform}{subterm types}

\item{Term_n}{Term numbers}

\item{totalnum}{total number of parameters}

\item{fir}{st term number}

\item{s_weights}{vector of weights for every row}

\item{T0}{Term value for each subterm}

\item{Td0}{Term by subterm derivative matrix}

\item{Tdd0}{Term by subterm second derivative matrix}

\item{Te}{temporary term storage matrix}

\item{R}{risk vector}

\item{Rd}{Risk first derivative matrix}

\item{Rdd}{Risk second derivative matrix}

\item{Dose}{term matrix}

\item{nonDose}{term matrix}

\item{TTerm}{Total term matrix}

\item{nonDose_LIN}{Linear term matrix}

\item{nonDose_PLIN}{Product linear term matrix}

\item{nonDose_LOGLIN}{Loglinear term matrix}

\item{RdR}{Risk to first derivative ratio matrix}

\item{RddR}{Risk to second derivative ratio matrix}

\item{nthreads}{number of threads available}

\item{debugging}{additional boolean for verbosity in testing}

\item{KeepConstant}{binary vector to identify constant parameters}

\item{gmix_theta}{theta value for geometric-mixture model}

\item{gmix_term}{list of 0/1 to identify which terms to set as Relative Risk (0) or Excess Risk (1)}
}
\value{
Updates matrices in place: Risk, Risk ratios
}
\description{
Utility function to calculate the risk and risk ratios with a weighting applied
\code{Make_Risks_Weighted} Called to update weighted risk matrices, Splits into cases based on model form, Uses lists of term numbers and types to apply different derivative formulas
}
