% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Make_Subterms}
\alias{Make_Subterms}
\title{Utility function to calculate the term and subterm values
\code{Make_Subterms} Called to update term matrices, Uses lists of term numbers and types to apply formulas}
\usage{
Make_Subterms(
  totalnum,
  Term_n,
  tform,
  dfc,
  fir,
  T0,
  Td0,
  Tdd0,
  Dose,
  nonDose,
  TTerm,
  nonDose_LIN,
  nonDose_PLIN,
  nonDose_LOGLIN,
  beta_0,
  df0,
  dint,
  nthreads,
  debugging
)
}
\arguments{
\item{totalnum}{Total number of terms}

\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{dfc}{covariate column numbers}

\item{fir}{first term number}

\item{T0}{Term by subterm matrix}

\item{Td0}{Term by subterm derivative matrix}

\item{Tdd0}{Term by subterm second derivative matrix}

\item{Dose}{Dose term matrix}

\item{nonDose}{nonDose term matrix}

\item{TTerm}{Total term matrix}

\item{nonDose_LIN}{Linear term matrix}

\item{nonDose_PLIN}{Product linear term matrix}

\item{nonDose_LOGLIN}{Loglinear term matrix}

\item{beta_0}{parameter list}

\item{df0}{covariate matrix}

\item{dint}{value used for threshold derivative finite step}

\item{nthreads}{number of threads to use}

\item{debugging}{debugging boolean}
}
\value{
Updates matrices in place: Sub-term matrices, Term matrices
}
\description{
Utility function to calculate the term and subterm values
\code{Make_Subterms} Called to update term matrices, Uses lists of term numbers and types to apply formulas
}
