% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Make_Subterms_Basic}
\alias{Make_Subterms_Basic}
\title{Utility function to calculate the term and subterm values with the basic model
\code{Make_Subterms_Basic} Called to update term matrices, Uses lists of term numbers and types to apply formulas}
\usage{
Make_Subterms_Basic(totalnum, dfc, T0, beta_0, df0, nthreads, debugging)
}
\arguments{
\item{totalnum}{Total number of terms}

\item{dfc}{covariate column numbers}

\item{T0}{subterm values}

\item{beta_0}{parameter list}

\item{df0}{covariate matrix}

\item{nthreads}{number of threads to use}

\item{debugging}{debugging boolean}
}
\value{
Updates matrices in place: Sub-term matrices, Term matrices
}
\description{
Utility function to calculate the term and subterm values with the basic model
\code{Make_Subterms_Basic} Called to update term matrices, Uses lists of term numbers and types to apply formulas
}
