% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mem_usage}
\alias{mem_usage}
\title{Utility function to find the memory usage
\code{mem_usage} Called within the code, takes the address to assign value to}
\usage{
mem_usage(vm_usage)
}
\arguments{
\item{vm_usage}{double to store peak usage at}
}
\value{
assigned peak memory value
}
\description{
Utility function to find the memory usage
\code{mem_usage} Called within the code, takes the address to assign value to
}
