% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Model_Results_Log}
\alias{Model_Results_Log}
\title{Saves information about a run to a log file}
\usage{
Model_Results_Log(
  log_file = "out.log",
  df = data.table(),
  out_list = list(),
  noprint = TRUE
)
}
\arguments{
\item{log_file}{file to save log to}

\item{df}{a data.table containing the columns of interest}

\item{out_list}{list output from a regression, used to build results table and pull out convergence values}

\item{noprint}{boolean, if true the file is saved to the log file, if false the output is printed to the console INSTEAD of being saved.}
}
\value{
null, prints to screen or saves to file
}
\description{
\code{Model_Results_Log} saves information about the data, results, model, computer, software, and date to an external file. Intended to make reproduction of results easier
}
\seealso{
Other Output and Information Functions: 
\code{\link{Interpret_Output}()},
\code{\link{System_Version}()}
}
\concept{Output and Information Functions}
