% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{nested_split}
\alias{nested_split}
\title{Splits a string by commas, skipping over paranthesis sections}
\usage{
nested_split(total_string)
}
\value{
returns a vector of substrings
}
\description{
\code{nested_split} splits by comma, keeps paranthesis sections together
}
\examples{
library(data.table)
## basic example code reproduced from the starting-description vignette
model <- 'd,factor(e,levels=c(1,2),labels=c("low","high")),f,g'
nested_split(model)
}
\seealso{
Other Data Cleaning Functions: 
\code{\link{Check_Dupe_Columns}()},
\code{\link{Check_Trunc}()},
\code{\link{Check_Verbose}()},
\code{\link{Date_Shift}()},
\code{\link{Def_Control}()},
\code{\link{Def_model_control}()},
\code{\link{Event_Count_Gen}()},
\code{\link{Event_Time_Gen}()},
\code{\link{Joint_Multiple_Events}()},
\code{\link{Replace_Missing}()},
\code{\link{Time_Since}()},
\code{\link{factorize}()},
\code{\link{factorize_par}()},
\code{\link{gen_time_dep}()},
\code{\link{interact_them}()},
\code{\link{parse_literal_string}()}
}
\concept{Data Cleaning Functions}
