% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OMP_Check}
\alias{OMP_Check}
\title{Checks the OMP flag}
\usage{
OMP_Check()
}
\value{
boolean: True for OMP allowed
}
\description{
\code{OMP_Check} Called directly from R, checks the omp flag and returns if omp is enabled
}
