% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Open_File}
\alias{Open_File}
\title{Opens a file for reading
\code{Open_File} uses user provided file name, time columns, and event column to load and sort.}
\usage{
Open_File(fname, time1, time2, event)
}
\arguments{
\item{fname}{a string file name}

\item{time1}{a column name defining the start of time periods, sorted third}

\item{time2}{a column name defining the end of time periods, sorted first}

\item{event}{a column name defining the events, sorted second}
}
\value{
returns a sorted dataframe
}
\description{
Opens a file for reading
\code{Open_File} uses user provided file name, time columns, and event column to load and sort.
}
