% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{parse_literal_string}
\alias{parse_literal_string}
\title{converts a string of a vector/list to vector/list}
\usage{
parse_literal_string(string)
}
\arguments{
\item{string}{the string to convert to vector/list}
}
\value{
returns a vector, list, or the original string
}
\description{
\code{parse_literal_string} converts the string of a vector/list to a vector/list
}
\seealso{
Other Data Cleaning Functions: 
\code{\link{Check_Dupe_Columns}()},
\code{\link{Check_Trunc}()},
\code{\link{Check_Verbose}()},
\code{\link{Date_Shift}()},
\code{\link{Def_Control}()},
\code{\link{Def_model_control}()},
\code{\link{Event_Count_Gen}()},
\code{\link{Event_Time_Gen}()},
\code{\link{Joint_Multiple_Events}()},
\code{\link{Replace_Missing}()},
\code{\link{Time_Since}()},
\code{\link{factorize}()},
\code{\link{factorize_par}()},
\code{\link{gen_time_dep}()},
\code{\link{interact_them}()},
\code{\link{nested_split}()}
}
\concept{Data Cleaning Functions}
