% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Plot_Omnibus}
\alias{Plot_Omnibus}
\title{Primary plotting function.
\code{Plot_Omnibus} Performs the calls to calculation functions}
\usage{
Plot_Omnibus(
  Term_n,
  tform,
  a_n,
  x_all,
  dfc,
  fir,
  der_iden,
  modelform,
  abs_max,
  dose_abs_max,
  df_groups,
  tu,
  verbose,
  debugging,
  KeepConstant,
  term_tot,
  ties_method,
  nthreads,
  STRATA_vals,
  cens_weight,
  cens_thres,
  uniq_v,
  strata_bool,
  basic_bool,
  CR_bool,
  Surv_bool,
  Risk_bool,
  Schoenfeld_bool,
  Risk_Sub_bool,
  gmix_theta,
  gmix_term
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{x_all}{covariate matrix}

\item{dfc}{vector matching subterm number to matrix column}

\item{fir}{st term number}

\item{der_iden}{subterm number for derivative tests}

\item{modelform}{string model identifier}

\item{abs_max}{Maximum allowed parameter change}

\item{dose_abs_max}{Maximum allowed threshold parameter change}

\item{df_groups}{matrix with time and event information}

\item{tu}{Event time vector}

\item{verbose}{boolean for additional printing}

\item{debugging}{additional boolean for verbosity in testing}

\item{KeepConstant}{binary vector to identify constant parameters}

\item{term_tot}{total number of terms}

\item{ties_method}{Ties method}

\item{nthreads}{number of threads available}

\item{STRATA_vals}{vector of strata identifier values}

\item{cens_weight}{vector of censoring weights}

\item{cens_thres}{threshold to add competing event to risk group}

\item{uniq_v}{number of unique covariate values}

\item{strata_bool}{boolean for stratification}

\item{basic_bool}{boolean for multiplicative log-linear model}

\item{CR_bool}{boolean for competing risks}

\item{Surv_bool}{boolean for competing risks}

\item{Risk_bool}{boolean for competing risks}

\item{Schoenfeld_bool}{boolean for competing risks}

\item{Risk_Sub_bool}{boolean for competing risks}

\item{gmix_theta}{theta value for geometric-mixture model}

\item{gmix_term}{list of 0/1 to identify which terms to set as Relative Risk (0) or Excess Risk (1)}
}
\value{
List of final results: Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, model information
}
\description{
Primary plotting function.
\code{Plot_Omnibus} Performs the calls to calculation functions
}
