% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Plot_Omnibus_transition}
\alias{Plot_Omnibus_transition}
\title{Interface between R code and the plotting omnibus function
\code{Plot_Omnibus_transition} Called directly from R, Defines the control variables and calls the plotting functions}
\usage{
Plot_Omnibus_transition(
  Term_n,
  tform,
  a_n,
  dfc,
  x_all,
  fir,
  der_iden,
  modelform,
  Control,
  df_groups,
  tu,
  KeepConstant,
  term_tot,
  STRATA_vals,
  cens_vec,
  model_control
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{dfc}{vector matching subterm number to matrix column}

\item{x_all}{covariate matrix}

\item{fir}{st term number}

\item{der_iden}{subterm number for derivative tests}

\item{modelform}{string model identifier}

\item{Control}{control list}

\item{df_groups}{matrix with time and event information}

\item{tu}{Event time vector}

\item{KeepConstant}{binary vector to identify constant parameters}

\item{term_tot}{total number of terms}

\item{STRATA_vals}{vector of strata identifier values}

\item{cens_vec}{censoring weight list}

\item{model_control}{controls which alternative model options are used}
}
\value{
LogLik_Cox_PH output : Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, model information
}
\description{
Interface between R code and the plotting omnibus function
\code{Plot_Omnibus_transition} Called directly from R, Defines the control variables and calls the plotting functions
}
