% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PLOT_SURV}
\alias{PLOT_SURV}
\title{Primary Cox PH baseline hazard function
\code{PLOT_SURV} Performs the calls to calculation functions, Uses calculated risks and risk groups to approximate the baseline, With verbose option prints out time stamps and intermediate sums of terms and derivatives}
\usage{
PLOT_SURV(reqrdnum, R, Rd, a_er, df_groups, tu, verbose, debugging, nthreads)
}
\arguments{
\item{reqrdnum}{total number of free parameters}

\item{R}{risk vector}

\item{Rd}{Risk first derivative matrix}

\item{a_er}{Optimal value standard error}

\item{df_groups}{matrix with time and event information}

\item{tu}{Event time vector}

\item{verbose}{boolean for additional printing}

\item{debugging}{additional boolean for verbosity in testing}

\item{nthreads}{number of threads available}
}
\value{
List of results: baseline hazard, risk for each row
}
\description{
Primary Cox PH baseline hazard function
\code{PLOT_SURV} Performs the calls to calculation functions, Uses calculated risks and risk groups to approximate the baseline, With verbose option prints out time stamps and intermediate sums of terms and derivatives
}
