% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTypes.R
\name{PlotCox_Schoenfeld_Residual}
\alias{PlotCox_Schoenfeld_Residual}
\title{Calculates Schoenfeld residuals for a Cox Proportional Hazards regression and plots}
\usage{
PlotCox_Schoenfeld_Residual(
  df,
  time1,
  time2,
  event0,
  names,
  Term_n,
  tform,
  keep_constant,
  a_n,
  modelform,
  fir,
  der_iden,
  control,
  age_unit,
  Plot_Name,
  model_control = list()
)
}
\arguments{
\item{df}{a data.table containing the columns of interest}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{event0}{column used for event status}

\item{names}{columns for elements of the model, used to identify data columns}

\item{Term_n}{term numbers for each element of the model}

\item{tform}{list of string function identifiers, used for linear/step}

\item{keep_constant}{binary values to denote which parameters to change}

\item{a_n}{list of initial parameter values, used to determine number of parameters}

\item{modelform}{string specifying the model type: M, ME, A, PA, PAE, GMIX, GMIX-R, GMIX-E}

\item{fir}{term number for the initial term, used for models of the form T0*f(Ti) in which the order matters}

\item{der_iden}{number for the subterm to test derivative at, only used for testing runs with a single varying parameter, should be smaller than total number of parameters}

\item{control}{list of parameters controlling the convergence, see Def_Control() for options or vignette("starting_description")}

\item{age_unit}{age unit}

\item{Plot_Name}{plot identifier, used in filename for saved plots}

\item{model_control}{controls which alternative model options are used, see Def_model_control() for options and vignette("Alt_Run_opt") for further details}
}
\value{
saves the plots in the current directory and returns a string that it passed
}
\description{
\code{RunCox_Schoenfeld_Residual} uses user provided data, time/event columns, vectors specifying the model, and options to calculate the residuals
}
\seealso{
Other {Plotting Functions}: 
\code{\link{CoxKaplanMeier}()},
\code{\link{CoxMartingale}()},
\code{\link{CoxRisk}()},
\code{\link{CoxStratifiedSurvival}()},
\code{\link{CoxSurvival}()},
\code{\link{GetCensWeight}()}
}
\concept{{Plotting Functions}}
