% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{plotMartingale.default}
\alias{plotMartingale.default}
\title{Generic Martingale Residual Plotting function, default option}
\usage{
\method{plotMartingale}{default}(x, df, plot_options, a_n = c(), ...)
}
\arguments{
\item{x}{result object from a regression, class coxres}

\item{df}{a data.table containing the columns of interest}

\item{plot_options}{list of parameters controlling the plot options, see RunCoxPlots() for different options}

\item{a_n}{list of initial parameter values, used to determine the number of parameters. May be either a list of vectors or a single vector.}

\item{...}{can include the named entries for the plot_options parameter}
}
\description{
\code{plotMartingale.default} Generic Martingale Residual Plotting, by default nothing happens
}
