% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_run.R
\name{plotSchoenfeld.coxres}
\alias{plotSchoenfeld.coxres}
\title{Performs Cox Proportional Hazard model schoenfeld residual plots}
\usage{
\method{plotSchoenfeld}{coxres}(x, df, plot_options, a_n = NULL, ...)
}
\arguments{
\item{x}{result object from a regression, class coxres}

\item{df}{a data.table containing the columns of interest}

\item{plot_options}{list of parameters controlling the plot options, see the \href{../articles/Plotting_And_Analysis.html}{plotting vignette} for different options}

\item{a_n}{list of initial parameter values, used to determine the number of parameters. May be either a list of vectors or a single vector.}

\item{...}{can include the named entries for the plot_options parameter}
}
\value{
returns the data used for plots
}
\description{
\code{plotSchoenfeld.coxres} uses user provided data, time/event columns,
vectors specifying the model, and options to choose and save plots
}
\seealso{
Other Plotting Wrapper Functions:
\code{\link[=plot.coxres]{plot.coxres()}},
\code{\link[=plotMartingale.coxres]{plotMartingale.coxres()}},
\code{\link[=plotRisk.coxres]{plotRisk.coxres()}},
\code{\link[=plotSurvival.coxres]{plotSurvival.coxres()}}
}
\concept{Plotting Wrapper Functions}
