% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_run.R
\name{plotSurvival}
\alias{plotSurvival}
\title{Generic Survival Plotting function}
\usage{
plotSurvival(x, df, plot_options, a_n = NULL, ...)
}
\arguments{
\item{x}{result object from a regression, class coxres}

\item{df}{a data.table containing the columns of interest}

\item{plot_options}{list of parameters controlling the plot options, see the \href{../articles/Plotting_And_Analysis.html}{plotting vignette} for different options}

\item{a_n}{list of initial parameter values, used to determine the number of parameters. May be either a list of vectors or a single vector.}

\item{...}{can include the named entries for the plot_options parameter}
}
\value{
returns the data used for plots
}
\description{
\code{plotSurvival} Generic survival plotting, used by the coxres
object. Calculates the baseline hazard, cumulative hazard, and survival
curve for the list of event times. Currently only returns data that could be
plotted.
}
