% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pois_Omnibus_transition}
\alias{pois_Omnibus_transition}
\title{Interface between R code and the pois omnibus regression function
\code{pois_Omnibus_transition} Called directly from R, Defines the control variables and calls the regression function}
\usage{
pois_Omnibus_transition(
  dfe,
  Term_n,
  tform,
  a_ns,
  dfc,
  x_all,
  fir,
  der_iden,
  modelform,
  Control,
  KeepConstant,
  term_tot,
  df0,
  model_control
)
}
\arguments{
\item{dfe}{Matrix with person-year/event count information}

\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_ns}{matrix of starting values}

\item{dfc}{covariate column numbers}

\item{x_all}{covariate matrix}

\item{fir}{first term number}

\item{der_iden}{subterm number for derivative tests}

\item{modelform}{model string}

\item{Control}{control list}

\item{KeepConstant}{vector of parameters to keep constant}

\item{term_tot}{total number of terms}

\item{df0}{matrix of strata identifier values}

\item{model_control}{controls which alternative model options are used}
}
\value{
LogLik_Cox_PH output : Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, model information
}
\description{
Interface between R code and the pois omnibus regression function
\code{pois_Omnibus_transition} Called directly from R, Defines the control variables and calls the regression function
}
