% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Poisson_LogLik_Single}
\alias{Poisson_LogLik_Single}
\title{Utility function to calculate poisson log-likelihood
\code{Poisson_LogLik_Single} Called to update log-likelihoods, Uses list risk matrices and person-years, Sums the log-likelihood contribution from each row}
\usage{
Poisson_LogLik_Single(nthreads, totalnum, PyrC, R, Ll, debugging)
}
\arguments{
\item{nthreads}{number of threads available}

\item{totalnum}{total number of parameters}

\item{PyrC}{matrix of person-years and event counts}

\item{R}{risk vector}

\item{Ll}{Log-likelihood vector}

\item{debugging}{additional boolean for verbosity in testing}
}
\value{
Updates matrices in place: Log-likelihood vectors/matrix
}
\description{
Utility function to calculate poisson log-likelihood
\code{Poisson_LogLik_Single} Called to update log-likelihoods, Uses list risk matrices and person-years, Sums the log-likelihood contribution from each row
}
