% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poisson_transition_guess}
\alias{poisson_transition_guess}
\title{Interface between R code and the poisson multi-start regression
\code{poisson_transition_guess} Called directly from R, Defines the control variables and calls the regression function}
\usage{
poisson_transition_guess(
  dfe,
  Term_n,
  tform,
  a_ns,
  dfc,
  x_all,
  fir,
  der_iden,
  modelform,
  Control,
  KeepConstant,
  term_tot
)
}
\arguments{
\item{dfe}{Matrix with person-year/event count information}

\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_ns}{matrix of starting values}

\item{dfc}{covariate column numbers}

\item{x_all}{covariate matrix}

\item{fir}{first term number}

\item{der_iden}{subterm number for derivative tests}

\item{modelform}{model string}

\item{Control}{control list}

\item{KeepConstant}{vector of parameters to keep constant}

\item{term_tot}{total number of terms}
}
\value{
LogLik_Cox_PH output : Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, model information
}
\description{
Interface between R code and the poisson multi-start regression
\code{poisson_transition_guess} Called directly from R, Defines the control variables and calls the regression function
}
