% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poisson_transition_single}
\alias{poisson_transition_single}
\title{Interface between R code and the poisson calculation
\code{poisson_transition_single} Called directly from R, Defines the control variables and calls the calculation function}
\usage{
poisson_transition_single(
  dfe,
  Term_n,
  tform,
  a_n,
  dfc,
  x_all,
  fir,
  modelform,
  Control,
  term_tot
)
}
\arguments{
\item{dfe}{Matrix with person-year/event count information}

\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{dfc}{covariate column numbers}

\item{x_all}{covariate matrix}

\item{fir}{first term number}

\item{modelform}{model string}

\item{Control}{control list}

\item{term_tot}{total number of terms}
}
\value{
LogLik_Poisson output : Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, deviance, model information
}
\description{
Interface between R code and the poisson calculation
\code{poisson_transition_single} Called directly from R, Defines the control variables and calls the calculation function
}
