% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Prep_Basic}
\alias{Prep_Basic}
\title{Utility function to calculate risks and derivatives for basic case
\code{Prep_Basic} Called to update term matrices, Uses lists of term numbers and types to apply formulas}
\usage{
Prep_Basic(
  totalnum,
  dfc,
  T0,
  Td0,
  Tdd0,
  beta_0,
  df0,
  nthreads,
  debugging,
  KeepConstant
)
}
\arguments{
\item{totalnum}{Total number of terms}

\item{dfc}{covariate column numbers}

\item{T0}{term values}

\item{Td0}{term derivative values}

\item{Tdd0}{term second derivatives values}

\item{beta_0}{parameter list}

\item{df0}{covariate matrix}

\item{nthreads}{number of threads to use}

\item{debugging}{debugging boolean}

\item{KeepConstant}{vector identifying constant parameters}
}
\value{
Updates matrices in place: Sub-term matrices, Term matrices
}
\description{
Utility function to calculate risks and derivatives for basic case
\code{Prep_Basic} Called to update term matrices, Uses lists of term numbers and types to apply formulas
}
