% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{print.coxres}
\alias{print.coxres}
\title{Prints a cox regression output clearly}
\usage{
\method{print}{coxres}(object, digits = 2)
}
\arguments{
\item{digits}{digits used for printing results}
}
\value{
return nothing, prints the results to console
}
\description{
\code{print.coxres} uses the list output from a regression, prints off a table of results and summarizes the score and convergence.
}
\seealso{
Other Output and Information Functions: 
\code{\link{Interpret_Output}()},
\code{\link{System_Version}()},
\code{\link{print.poisres}()}
}
\concept{Output and Information Functions}
