% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{print.logitresbound}
\alias{print.logitresbound}
\title{Prints a logistic likelihood boundary regression output clearly}
\usage{
\method{print}{logitresbound}(x, ...)
}
\arguments{
\item{x}{result object from a regression, class logitresbound}

\item{...}{can include the number of digits, named digit, or an unnamed integer entry assumed to be digits}
}
\value{
return nothing, prints the results to console
}
\description{
\code{print.logitresbound} uses the list output from a regression, prints off a table of results and summarizes the score and convergence.
}
\seealso{
Other Output and Information Functions:
\code{\link[=System_Version]{System_Version()}},
\code{\link[=print.caseconres]{print.caseconres()}},
\code{\link[=print.coxres]{print.coxres()}},
\code{\link[=print.coxresbound]{print.coxresbound()}},
\code{\link[=print.logitres]{print.logitres()}},
\code{\link[=print.poisres]{print.poisres()}},
\code{\link[=print.poisresbound]{print.poisresbound()}}
}
\concept{Output and Information Functions}
