% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Rcomp_version}
\alias{Rcomp_version}
\title{Checks how R was compiled}
\usage{
Rcomp_version()
}
\value{
returns a string representation of gcc, clang, or R CMD config CC output
}
\description{
\code{Rcomp_version} Checks how R was compiled, part of configuration script
}
