% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Rcpp_version}
\alias{Rcpp_version}
\title{Checks default R c++ compiler}
\usage{
Rcpp_version()
}
\value{
returns a string representation of gcc, clang, or head ~/.R/Makevars
}
\description{
\code{Rcpp_version} checks ~/.R/Makevars script for default compilers set, part of configuration script
}
