% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{RelativeRisk.default}
\alias{RelativeRisk.default}
\title{Generic relative risk calculation function, default option}
\usage{
\method{RelativeRisk}{default}(x, df, ...)
}
\arguments{
\item{x}{result object from a regression, class coxres}

\item{df}{a data.table containing the columns of interest}

\item{...}{extended for other necessary parameters}
}
\description{
\code{RelativeRisk.default} Generic relative risk calculation function, by default nothing happens
}
