% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{removeColumn}
\alias{removeColumn}
\title{Utility function to remove columns
\code{removeColumn} Called within the code, removes and resizes the matrix}
\usage{
removeColumn(matrix_modify, colToRemove)
}
\arguments{
\item{matrix_modify}{matrix to remove rows or columns from}

\item{colToRemove}{column index to remove}
}
\value{
eigen matrix with the column removed
}
\description{
Utility function to remove columns
\code{removeColumn} Called within the code, removes and resizes the matrix
}
