% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{removeRow}
\alias{removeRow}
\title{Utility function to remove rows
\code{removeRow} Called within the code, removes and resizes the matrix}
\usage{
removeRow(matrix, rowToRemove)
}
\arguments{
\item{matrix}{Eigen matrix to change}

\item{rowToRemove}{row index to remove}
}
\value{
eigen matrix with the row removed
}
\description{
Utility function to remove rows
\code{removeRow} Called within the code, removes and resizes the matrix
}
