% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{Residual.default}
\alias{Residual.default}
\title{Generic Residual calculation function, default option}
\usage{
\method{Residual}{default}(x, df, ...)
}
\arguments{
\item{x}{result object from a regression, class coxres or poisres}

\item{df}{a data.table containing the columns of interest}

\item{...}{extended for other necessary parameters}
}
\description{
\code{Residual.default} Generic Residual calculation function, by default nothing happens
}
