% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_run.R
\name{Residual.logitres}
\alias{Residual.logitres}
\title{Calculates the Residuals for a completed logistic model}
\usage{
\method{Residual}{logitres}(
  x,
  df,
  control = list(),
  a_n = NULL,
  pearson = FALSE,
  deviance = FALSE,
  ...
)
}
\arguments{
\item{x}{result object from a regression, class logitres}

\item{df}{a data.table containing the columns of interest}

\item{control}{list of parameters controlling the convergence, see the \href{../articles/Control_Options.html}{control options vignette} for details}

\item{a_n}{list of initial parameter values, used to determine the number of parameters. May be either a list of vectors or a single vector.}

\item{pearson}{boolean to calculate pearson residuals}

\item{deviance}{boolean to calculate deviance residuals}

\item{...}{can include the named entries for the assign_control list parameter}
}
\value{
returns matrices of residuals
}
\description{
\code{Residual.logitres} uses user provided data, person-year/event columns, vectors specifying the model,
and options to calculate residuals for a solved Poisson regression
}
\seealso{
Other Residuals:
\code{\link[=Residual]{Residual()}},
\code{\link[=Residual.logitmodel]{Residual.logitmodel()}},
\code{\link[=Residual.poismodel]{Residual.poismodel()}},
\code{\link[=Residual.poisres]{Residual.poisres()}}
}
\concept{Residuals}
