% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RISK_SUBSET}
\alias{RISK_SUBSET}
\title{Primary reference vector risk function
\code{RISK_SUBSET} Performs the calls to calculation functions, Calculates risk for a set of reference vectors, With verbose option prints out time stamps and intermediate sums of terms and derivatives}
\usage{
RISK_SUBSET(
  Term_n,
  tform,
  a_n,
  x_all,
  dfc,
  fir,
  modelform,
  verbose,
  debugging,
  term_tot,
  nthreads
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{x_all}{covariate matrix}

\item{dfc}{covariate column numbers}

\item{fir}{first term number}

\item{modelform}{model string}

\item{verbose}{verbosity boolean}

\item{debugging}{debugging boolean}

\item{term_tot}{total number of terms}

\item{nthreads}{number of threads to use}
}
\value{
List of results: Risk at the reference
}
\description{
Primary reference vector risk function
\code{RISK_SUBSET} Performs the calls to calculation functions, Calculates risk for a set of reference vectors, With verbose option prints out time stamps and intermediate sums of terms and derivatives
}
