% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cox_Regression.R
\name{RunCoxNull}
\alias{RunCoxNull}
\title{Performs basic Cox Proportional Hazards regression with the null model
\code{RunCoxRegression} uses user provided data and time/event columns to calculate the log-likelihood with constant hazard ratio}
\usage{
RunCoxNull(df, time1, time2, event, control)
}
\arguments{
\item{df}{data used for regression}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{event}{column used for event status}

\item{control}{list of parameters controlling the convergence}
}
\value{
returns a list of the final results
}
\description{
Performs basic Cox Proportional Hazards regression with the null model
\code{RunCoxRegression} uses user provided data and time/event columns to calculate the log-likelihood with constant hazard ratio
}
