% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cox_Regression.R
\name{RunCoxPlots}
\alias{RunCoxPlots}
\title{Performs Cox Proportional Hazard model plots}
\usage{
RunCoxPlots(
  df,
  time1 = "\%trunc\%",
  time2 = "\%trunc\%",
  event0 = "event",
  names = c("CONST"),
  term_n = c(0),
  tform = "loglin",
  keep_constant = c(0),
  a_n = c(0),
  modelform = "M",
  control = list(),
  plot_options = list(),
  model_control = list()
)
}
\arguments{
\item{df}{a data.table containing the columns of interest}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{event0}{column used for event status}

\item{names}{columns for elements of the model, used to identify data columns}

\item{term_n}{term numbers for each element of the model}

\item{tform}{list of string function identifiers, used for linear/step}

\item{keep_constant}{binary values to denote which parameters to change}

\item{a_n}{list of initial parameter values, used to determine the number of parameters. May be either a list of vectors or a single vector.}

\item{modelform}{string specifying the model type: M, ME, A, PA, PAE, GMIX, GMIX-R, GMIX-E}

\item{control}{list of parameters controlling the convergence, see the Control_Options vignette for details}

\item{plot_options}{list of parameters controlling the plot options, see RunCoxPlots() for different options}

\item{model_control}{controls which alternative model options are used, see the Control_Options vignette for further details}
}
\value{
saves the plots in the current directory and returns the data used for plots
}
\description{
\code{RunCoxPlots} uses user provided data, time/event columns,
vectors specifying the model, and options to choose and save plots
}
\seealso{
Other Plotting Wrapper Functions: 
\code{\link{Cox_Relative_Risk}()},
\code{\link{plot.coxres}()}
}
\concept{Plotting Wrapper Functions}
