% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cox_Regression.R
\name{RunCoxRegression_Basic}
\alias{RunCoxRegression_Basic}
\title{Performs basic Cox Proportional Hazards regression with a basic model
\code{RunCoxRegression_Basic} uses user provided data, time/event columns, vectors specifying the model, and options to control the convergence and starting positions}
\usage{
RunCoxRegression_Basic(
  df,
  time1,
  time2,
  event,
  names,
  keep_constant,
  a_n,
  der_iden,
  control
)
}
\arguments{
\item{df}{data used for regression}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{event}{column used for event status}

\item{names}{columns names for elements of the model, used to identify data columns}

\item{keep_constant}{vector of 0/1 to identify parameters to force to be constant}

\item{a_n}{starting parameters for regression}

\item{der_iden}{number for the subterm to test derivative at, only used for testing runs with a single varying parameter}

\item{control}{list of parameters controlling the convergence}
}
\value{
returns a list of the final results
}
\description{
Performs basic Cox Proportional Hazards regression with a basic model
\code{RunCoxRegression_Basic} uses user provided data, time/event columns, vectors specifying the model, and options to control the convergence and starting positions
}
