% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_Regression.R
\name{RunPoisRegression_Omnibus_Multidose}
\alias{RunPoisRegression_Omnibus_Multidose}
\title{Performs Poisson regression using the omnibus function with multiple column realizations}
\usage{
RunPoisRegression_Omnibus_Multidose(
  df,
  pyr0 = "pyr",
  event0 = "event",
  names = c("CONST"),
  term_n = c(0),
  tform = "loglin",
  keep_constant = c(0),
  a_n = c(0),
  modelform = "M",
  realization_columns = matrix(c("temp00", "temp01", "temp10", "temp11"), nrow = 2),
  realization_index = c("temp0", "temp1"),
  control = list(),
  strat_col = "null",
  model_control = list(),
  cons_mat = as.matrix(c(0)),
  cons_vec = c(0)
)
}
\arguments{
\item{df}{a data.table containing the columns of interest}

\item{pyr0}{column used for person-years per row}

\item{event0}{column used for event status}

\item{names}{columns for elements of the model, used to identify data columns}

\item{term_n}{term numbers for each element of the model}

\item{tform}{list of string function identifiers, used for linear/step}

\item{keep_constant}{binary values to denote which parameters to change}

\item{a_n}{list of initial parameter values, used to determine the number of parameters. May be either a list of vectors or a single vector.}

\item{modelform}{string specifying the model type: M, ME, A, PA, PAE, GMIX, GMIX-R, GMIX-E}

\item{realization_columns}{used for multi-realization regressions. Matrix of column names with rows for each column with realizations, columns for each realization}

\item{realization_index}{used for multi-realization regressions. Vector of column names, one for each column with realizations. Each name should be used in the "names" variable in the equation definition}

\item{control}{list of parameters controlling the convergence, see the Control_Options vignette for details}

\item{strat_col}{column to stratify by if needed}

\item{model_control}{controls which alternative model options are used, see the Control_Options vignette for further details}

\item{cons_mat}{Matrix containing coefficients for a system of linear constraints, formatted as matrix}

\item{cons_vec}{Vector containing constants for a system of linear constraints, formatted as vector}
}
\value{
returns a list of the final results for each realization
}
\description{
\code{RunPoisRegression_Omnibus_Multidose} uses user provided data, time/event columns,
vectors specifying the model, and options to control the convergence
and starting positions. Used for 2DMC column uncertainty methods.
Returns optimized parameters, log-likelihood, and standard deviation for each realization.
Has additional options for using stratification
}
\seealso{
Other Poisson Wrapper Functions: 
\code{\link{EventAssignment.poisres}()},
\code{\link{LikelihoodBound.poisres}()},
\code{\link{PoisRun}()},
\code{\link{PoisRunJoint}()},
\code{\link{PoisRunMulti}()},
\code{\link{PoissonCurveSolver}()},
\code{\link{Residual.poisres}()},
\code{\link{RunPoissonEventAssignment}()},
\code{\link{RunPoissonEventAssignment_bound}()},
\code{\link{RunPoissonRegression_Joint_Omnibus}()},
\code{\link{RunPoissonRegression_Omnibus}()},
\code{\link{RunPoissonRegression_Residual}()}
}
\concept{Poisson Wrapper Functions}
