% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_Regression.R
\name{RunPoissonEventAssignment_bound}
\alias{RunPoissonEventAssignment_bound}
\title{Predicts how many events are due to baseline vs excess at the confidence bounds of a single parameter}
\usage{
RunPoissonEventAssignment_bound(
  df,
  pyr0 = "pyr",
  event0 = "event",
  alternative_model = list(),
  keep_constant = c(0),
  modelform = "M",
  check_num = 1,
  z = 2,
  control = list(),
  strat_col = "null",
  model_control = list()
)
}
\arguments{
\item{df}{a data.table containing the columns of interest}

\item{pyr0}{column used for person-years per row}

\item{event0}{column used for event status}

\item{alternative_model}{the new model of interest in list form, output from a Poisson regression}

\item{keep_constant}{binary values to denote which parameters to change}

\item{modelform}{string specifying the model type: M, ME, A, PA, PAE, GMIX, GMIX-R, GMIX-E}

\item{check_num}{the parameter number to check at the bounds of, indexed from 1 using the order returned by Colossus}

\item{z}{Z score to use for confidence interval}

\item{control}{list of parameters controlling the convergence, see the Control_Options vignette for details}

\item{strat_col}{column to stratify by if needed}

\item{model_control}{controls which alternative model options are used, see the Control_Options vignette for further details}
}
\value{
returns a list of the final results
}
\description{
\code{RunPoissonEventAssignment_bound} uses user provided data, the results of a poisson regression, and options to calculate background and excess events
}
\seealso{
Other Poisson Wrapper Functions: 
\code{\link{EventAssignment.poisres}()},
\code{\link{LikelihoodBound.poisres}()},
\code{\link{PoisRun}()},
\code{\link{PoisRunJoint}()},
\code{\link{PoisRunMulti}()},
\code{\link{PoissonCurveSolver}()},
\code{\link{Residual.poisres}()},
\code{\link{RunPoisRegression_Omnibus_Multidose}()},
\code{\link{RunPoissonEventAssignment}()},
\code{\link{RunPoissonRegression_Joint_Omnibus}()},
\code{\link{RunPoissonRegression_Omnibus}()},
\code{\link{RunPoissonRegression_Residual}()}
}
\concept{Poisson Wrapper Functions}
