% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_Regression.R
\name{RunPoissonRegression}
\alias{RunPoissonRegression}
\title{Performs basic poisson regression
\code{RunPoissonRegression} uses user provided data, person-year/event columns, vectors specifying the model, and options to control the convergence and starting positions}
\usage{
RunPoissonRegression(
  df,
  time1,
  time2,
  pyr,
  event,
  names,
  Term_n,
  tform,
  keep_constant,
  a_n,
  modelform,
  fir,
  der_iden,
  control
)
}
\arguments{
\item{df}{data used for regression}

\item{time1}{column used for time period starts, may not be needed}

\item{time2}{column used for time period end, may not be needed}

\item{pyr}{column used for person-years per row}

\item{event}{column used for event status}

\item{names}{columns names for elements of the model, used to identify data columns}

\item{Term_n}{term numbers for each element of the model}

\item{tform}{subterm type for each element of the model}

\item{keep_constant}{vector of 0/1 to identify parameters to force to be constant}

\item{a_n}{starting parameters for regression}

\item{modelform}{string specifying the model type}

\item{fir}{term number for the intial term, used for models of the form T0*f(Ti) in which the order matters}

\item{der_iden}{number for the subterm to test derivative at, only used for testing runs with a single varying parameter}

\item{control}{list of parameters controlling the convergence}
}
\value{
prints the final results and return null
}
\description{
Performs basic poisson regression
\code{RunPoissonRegression} uses user provided data, person-year/event columns, vectors specifying the model, and options to control the convergence and starting positions
}
