% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_Regression.R
\name{RunPoissonRegression_Tier_Guesses}
\alias{RunPoissonRegression_Tier_Guesses}
\title{Performs basic poisson regression, with multiple guesses, starts with a single term
\code{RunPoissonRegression_Tier_Guesses} uses user provided data, time/event columns, vectors specifying the model, and options to control the convergence and starting positions, with additional guesses}
\usage{
RunPoissonRegression_Tier_Guesses(
  df,
  pyr,
  event,
  names,
  Term_n,
  tform,
  keep_constant,
  a_n,
  modelform,
  fir,
  der_iden,
  control,
  Strat_Cols,
  guesses_control
)
}
\arguments{
\item{df}{data used for regression}

\item{pyr}{column used for person-years per row}

\item{event}{column used for event status}

\item{names}{columns names for elements of the model, used to identify data columns}

\item{Term_n}{term numbers for each element of the model}

\item{tform}{subterm type for each element of the model}

\item{keep_constant}{vector of 0/1 to identify parameters to force to be constant}

\item{a_n}{starting parameters for regression}

\item{modelform}{string specifying the model type}

\item{fir}{term number for the intial term, used for models of the form T0*f(Ti) in which the order matters}

\item{der_iden}{number for the subterm to test derivative at, only used for testing runs with a single varying parameter}

\item{control}{list of parameters controlling the convergence}

\item{Strat_Cols}{column to stratify by}

\item{guesses_control}{list of parameters to control how the guessing works}
}
\value{
returns a list of the final results
}
\description{
Performs basic poisson regression, with multiple guesses, starts with a single term
\code{RunPoissonRegression_Tier_Guesses} uses user provided data, time/event columns, vectors specifying the model, and options to control the convergence and starting positions, with additional guesses
}
