% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Schoenfeld_Cox_PH}
\alias{Schoenfeld_Cox_PH}
\title{Primary Cox PH schoenfeld residual function
\code{Schoenfeld_Cox_PH} Performs the calls to calculation functions, Uses calculated risks and risk groups to calculate the residuals, With verbose option prints out time stamps and intermediate sums of terms and derivatives}
\usage{
Schoenfeld_Cox_PH(
  Term_n,
  tform,
  a_n,
  x_all,
  dfc,
  fir,
  der_iden,
  modelform,
  abs_max,
  dose_abs_max,
  df_groups,
  tu,
  verbose,
  debugging,
  KeepConstant,
  term_tot,
  ties_method
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{x_all}{covariate matrix}

\item{dfc}{covariate column numbers}

\item{fir}{first term number}

\item{der_iden}{subterm number for derivative tests}

\item{modelform}{model string}

\item{abs_max}{Maximum allowed parameter change}

\item{dose_abs_max}{Maximum allowed threshold parameter change}

\item{df_groups}{matrix with time and event information}

\item{tu}{event times}

\item{verbose}{verbosity boolean}

\item{debugging}{debugging boolean}

\item{KeepConstant}{vector identifying constant parameters}

\item{term_tot}{total number of terms}

\item{ties_method}{ties method}
}
\value{
List of results: scaled schoenfeld residuals
}
\description{
Primary Cox PH schoenfeld residual function
\code{Schoenfeld_Cox_PH} Performs the calls to calculation functions, Uses calculated risks and risk groups to calculate the residuals, With verbose option prints out time stamps and intermediate sums of terms and derivatives
}
