% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Stress_Test}
\alias{Stress_Test}
\title{Interface between R code and the Cox PH stress tester
\code{Stress_Test} Called directly from R, Defines the verbosity and tie method variables, Calls the calculation function}
\usage{
Stress_Test(
  Term_n,
  tform,
  a_n,
  dfc,
  x_all,
  fir,
  der_iden,
  modelform,
  Control,
  df_groups,
  tu,
  KeepConstant,
  term_tot,
  test_point
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{dfc}{covariate column numbers}

\item{x_all}{covariate matrix}

\item{fir}{first term number}

\item{der_iden}{subterm number for derivative tests}

\item{modelform}{model string}

\item{Control}{control list}

\item{df_groups}{time and event matrix}

\item{tu}{event times}

\item{KeepConstant}{vector of parameters to keep constant}

\item{term_tot}{total number of terms}

\item{test_point}{string vector of functions to test further}
}
\description{
Interface between R code and the Cox PH stress tester
\code{Stress_Test} Called directly from R, Defines the verbosity and tie method variables, Calls the calculation function
}
