% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{System_Version}
\alias{System_Version}
\title{Checks OS, compilers, and OMP}
\usage{
System_Version()
}
\value{
returns a list of results
}
\description{
\code{System_Version} checks OS, default R c++ compiler, and if OMP is enabled
}
\examples{
System_Version()
}
\seealso{
Other Output and Information Functions:
\code{\link[=OMP_Check]{OMP_Check()}}
}
\concept{Output and Information Functions}
