% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{time_factor}
\alias{time_factor}
\title{Defines Time Dependent Parameters
\code{time_factor} uses user provided bins and a list of columns to define interaction terms and update the data.table.
Technically could be used to define interaction terms for a binned column and a list of non-binned values}
\usage{
time_factor(df, time_bins, col_list, time_col)
}
\arguments{
\item{df}{a data.table containing the columns of interest}

\item{time_bins}{an array of bins to apply to the time column to split by}

\item{col_list}{an array of column names that should have interaction terms defined}

\item{time_col}{the column name of the time column the data is binned by}
}
\value{
returns a list with two named fields. df for the updated dataframe, and cols for the new column names
}
\description{
Defines Time Dependent Parameters
\code{time_factor} uses user provided bins and a list of columns to define interaction terms and update the data.table.
Technically could be used to define interaction terms for a binned column and a list of non-binned values
}
