% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_norm}
\alias{vec_norm}
\title{Utility function to take the norm of a vector
\code{vec_norm} Called within the code, takes a vector and length and finds the norm}
\usage{
vec_norm(x, totalnum)
}
\arguments{
\item{x}{std::vector to take norm of, assumed doubles}

\item{totalnum}{vector length}
}
\value{
double of the norm
}
\description{
Utility function to take the norm of a vector
\code{vec_norm} Called within the code, takes a vector and length and finds the norm
}
