% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Check_Dupe_Columns}
\alias{Check_Dupe_Columns}
\title{checks for duplicated column names
\code{Check_Dupe_Columns} checks for duplicated columns, columns with the same values, and columns with 1 value}
\usage{
Check_Dupe_Columns(df, cols, verbose)
}
\arguments{
\item{df}{dataframe of data to use as reference}

\item{cols}{columns to check}

\item{verbose}{verbosity}
}
\value{
returns the usable columns
}
\description{
checks for duplicated column names
\code{Check_Dupe_Columns} checks for duplicated columns, columns with the same values, and columns with 1 value
}
