% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Check_Risk}
\alias{Check_Risk}
\title{checks if the model is viable
\code{Check_Risk} Calculates risks and checks for negative values}
\usage{
Check_Risk(
  Term_n,
  tform,
  a_n,
  x_all,
  dfc,
  fir,
  modelform,
  verbose,
  debugging,
  KeepConstant,
  term_tot,
  nthreads,
  gmix_theta,
  gmix_term
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{x_all}{covariate matrix}

\item{dfc}{vector matching subterm number to matrix column}

\item{fir}{st term number}

\item{modelform}{string model identifier}

\item{verbose}{boolean for additional printing}

\item{debugging}{additional boolean for verbosity in testing}

\item{KeepConstant}{binary vector to identify constant parameters}

\item{term_tot}{total number of terms}

\item{nthreads}{number of threads available}

\item{gmix_theta}{theta value for geometric-mixture model}

\item{gmix_term}{list of 0/1 to identify which terms to set as Relative Risk (0) or Excess Risk (1)}
}
\value{
True for viable point, False for negative error
}
\description{
checks if the model is viable
\code{Check_Risk} Calculates risks and checks for negative values
}
