% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Check_Strata_Model}
\alias{Check_Strata_Model}
\title{Checks the default value for a given model, if every parameter were 0}
\usage{
Check_Strata_Model(term_n, tform, modelform, gmix_term, gmix_theta)
}
\arguments{
\item{term_n}{term numbers for each element of the model}

\item{tform}{list of string function identifiers, used for linear/step}

\item{modelform}{string specifying the model type: M, ME, A, PA, PAE, GMIX, GMIX-R, GMIX-E}

\item{gmix_term}{binary vector to denote excess (1) and relative terms (0). Excess terms have 1 added before use in risk model}

\item{gmix_theta}{double, used in gmix model. Multiplicative combination is taken to power of theta, additive combination is taken to 1-theta.}
}
\value{
TRUE if passed
}
\description{
\code{Check_Strata_Model} checks if a model is valid for stratified poisson
}
\seealso{
Other Data Cleaning Functions: 
\code{\link{Check_Iters}()},
\code{\link{Date_Shift}()},
\code{\link{Event_Count_Gen}()},
\code{\link{Event_Time_Gen}()},
\code{\link{Joint_Multiple_Events}()},
\code{\link{Replace_Missing}()},
\code{\link{Time_Since}()},
\code{\link{apply_norm}()},
\code{\link{factorize}()},
\code{\link{gen_time_dep}()}
}
\concept{Data Cleaning Functions}
