% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Check_Trunc}
\alias{Check_Trunc}
\title{Applies time duration truncation limits
\code{Check_Trunc} creates columns to use for truncation}
\usage{
Check_Trunc(df, ce, verbose = FALSE)
}
\arguments{
\item{df}{dataframe of data to use as reference}

\item{ce}{columns to check for truncation}

\item{verbose}{verbosity argument, for error returns}
}
\value{
returns the updated data and time period columns
}
\description{
Applies time duration truncation limits
\code{Check_Trunc} creates columns to use for truncation
}
