% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{ColossusCoxStrataSurv}
\alias{ColossusCoxStrataSurv}
\title{Interprets basic cox survival formula RHS with strata}
\usage{
ColossusCoxStrataSurv(...)
}
\value{
returns list with interval endpoints, event, and strata column
}
\description{
\code{ColossusCoxStrataSurv} assigns and interprets interval columns for cox model with stratification
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCoxSurv}()},
\code{\link{ColossusFineGrayStrataSurv}()},
\code{\link{ColossusFineGraySurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()}
}
\concept{Formula Interpretation}
