% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{ColossusCoxSurv}
\alias{ColossusCoxSurv}
\title{Interprets basic cox survival formula RHS}
\usage{
ColossusCoxSurv(...)
}
\value{
returns list with interval endpoints and event
}
\description{
\code{ColossusCoxSurv} assigns and interprets interval columns for cox model
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCoxStrataSurv}()},
\code{\link{ColossusFineGrayStrataSurv}()},
\code{\link{ColossusFineGraySurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()},
\code{\link{get_form_list}()},
\code{\link{get_form_risk}()},
\code{\link{get_form_surv}()}
}
\concept{Formula Interpretation}
