% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{ColossusFineGraySurv}
\alias{ColossusFineGraySurv}
\title{Interprets basic fine-gray survival formula RHS}
\usage{
ColossusFineGraySurv(...)
}
\arguments{
\item{...}{entries for a Fine_Gray object, tstart, tend, event, and weighting. Either in order or named. If unnamed, the last is assumed to be weighting and the rest are passed to ColossusCoxSurv.}
}
\value{
returns list with interval endpoints, event, and weighting columns
}
\description{
\code{ColossusFineGraySurv} assigns and interprets interval columns for fine-gray model
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCaseConStrataSurv}()},
\code{\link{ColossusCaseConSurv}()},
\code{\link{ColossusCaseConTimeStrataSurv}()},
\code{\link{ColossusCaseConTimeSurv}()},
\code{\link{ColossusCoxStrataSurv}()},
\code{\link{ColossusCoxSurv}()},
\code{\link{ColossusFineGrayStrataSurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()},
\code{\link{get_form_joint}()},
\code{\link{get_form_list}()},
\code{\link{get_form_risk}()},
\code{\link{get_form_surv}()}
}
\concept{Formula Interpretation}
