% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{ColossusLogitSurv}
\alias{ColossusLogitSurv}
\title{Interprets basic logistic survival formula RHS with no grouping}
\usage{
ColossusLogitSurv(...)
}
\arguments{
\item{...}{entries for a Logistic object, trials and events. trials not provided assumes one trial per row.}
}
\value{
returns list with event
}
\description{
\code{ColossusLogitSurv} assigns and interprets columns for trials and events in logistic model with no grouping.
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCoxSurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()},
\code{\link{get_form_joint}()}
}
\concept{Formula Interpretation}
