% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cox_Regression.R
\name{Cox_Relative_Risk}
\alias{Cox_Relative_Risk}
\title{Calculates proportional hazard for a reference vector
\code{RunCoxRegression} uses user provided data, time/event columns, vectors specifying the model, and options to calculate risk for a reference}
\usage{
Cox_Relative_Risk(
  df,
  time1 = "age_start",
  time2 = "age_exit",
  event = "cases",
  names = c("dose"),
  Term_n = rep(0, length(names)),
  tform = rep("loglin", length(names)),
  keep_constant = rep(0, length(names)),
  a_n = rep(0.01, length(names)),
  modelform = "M",
  fir = 0,
  control = list(lr = 0.75, maxiter = 20, halfmax = 5, epsilon = 1e-09, dbeta_max = 0.5,
    deriv_epsilon = 1e-09, abs_max = 1, change_all = TRUE, dose_abs_max = 100, verbose =
    FALSE, ties = "breslow", double_step = 1)
)
}
\arguments{
\item{df}{data used to calculate PH}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{event}{column used for event status}

\item{names}{columns names for elements of the model, used to identify data columns}

\item{Term_n}{term numbers for each element of the model}

\item{tform}{subterm type for each element of the model}

\item{keep_constant}{vector of 0/1 to identify parameters to force to be constant}

\item{a_n}{starting parameters for regression}

\item{modelform}{string specifying the model type}

\item{fir}{term number for the intial term, used for models of the form T0*f(Ti) in which the order matters}

\item{control}{list of parameters for verbosity and tie method}
}
\value{
returns a list of the final results
}
\description{
Calculates proportional hazard for a reference vector
\code{RunCoxRegression} uses user provided data, time/event columns, vectors specifying the model, and options to calculate risk for a reference
}
