% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTypes.r
\name{CoxMartingale}
\alias{CoxMartingale}
\title{calculates and plots martingale residuals with a named dose column
\code{CoxMartingale} uses user provided data, columns, and identifier to create plots}
\usage{
CoxMartingale(
  verbose,
  df,
  time1,
  time2,
  event,
  e,
  t,
  ch,
  dnames,
  Plot_Name,
  age_unit
)
}
\arguments{
\item{verbose}{boolean controlling additional printing}

\item{df}{data with covariates and times}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{event}{column used for event status}

\item{e}{output from a baseline calculation}

\item{t}{event times}

\item{ch}{cumulative hazards}

\item{dnames}{list of covariate columns to plot by}

\item{Plot_Name}{plot identifier}

\item{age_unit}{age unit}
}
\value{
saves the plots in the current directory and returns a string that it passed
}
\description{
calculates and plots martingale residuals with a named dose column
\code{CoxMartingale} uses user provided data, columns, and identifier to create plots
}
