% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTypes.R
\name{CoxMartingale}
\alias{CoxMartingale}
\title{calculates and plots martingale residuals with a named dose column
\code{CoxMartingale} uses user provided data, columns, and identifier to create plots}
\usage{
CoxMartingale(
  verbose,
  df,
  time1,
  time2,
  event0,
  e,
  t,
  ch,
  dnames,
  Plot_Name,
  age_unit,
  studyID
)
}
\arguments{
\item{verbose}{boolean to control if additional information is printed to the console}

\item{df}{a data.table containing the columns of interest}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{event0}{column used for event status}

\item{e}{output from a baseline calculation}

\item{t}{event times}

\item{ch}{cumulative hazards of baseline}

\item{dnames}{list of covariate columns to plot by}

\item{Plot_Name}{plot identifier}

\item{age_unit}{age unit}

\item{studyID}{id to group by, NaN for no grouping}
}
\value{
saves the plots in the current directory and returns a string that it passed
}
\description{
calculates and plots martingale residuals with a named dose column
\code{CoxMartingale} uses user provided data, columns, and identifier to create plots
}
