% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTypes.R
\name{CoxStratifiedSurvival}
\alias{CoxStratifiedSurvival}
\title{calculates and plots survival curves for each unique value of the stratification column
\code{CoxStratifiedSurvival} uses user provided data, columns, and identifier to calculate the survival fraction for each strata}
\usage{
CoxStratifiedSurvival(
  verbose,
  df,
  event0,
  time1,
  time2,
  names,
  Term_n,
  tform,
  a_n,
  er,
  fir,
  der_iden,
  modelform,
  control,
  keep_constant,
  Plot_Type,
  Strat_Col,
  time_lims,
  age_unit,
  model_control = list()
)
}
\arguments{
\item{verbose}{boolean to control if additional information is printed to the console, also accepts 0/1 integer}

\item{df}{a data.table containing the columns of interest}

\item{event0}{column used for event status}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{names}{columns for elements of the model, used to identify data columns}

\item{Term_n}{term numbers for each element of the model}

\item{tform}{list of string function identifiers, used for linear/step}

\item{a_n}{list of initial parameter values, used to determine number of parameters}

\item{er}{standard deviation for the parameters}

\item{fir}{term number for the initial term, used for models of the form T0*f(Ti) in which the order matters}

\item{der_iden}{number for the subterm to test derivative at, only used for testing runs with a single varying parameter, should be smaller than total number of parameters}

\item{modelform}{string specifying the model type: M, ME, A, PA, PAE, GMIX, GMIX-R, GMIX-E}

\item{control}{list of parameters controlling the convergence, see Def_Control() for options}

\item{keep_constant}{binary values to denote which parameters to change}

\item{Plot_Type}{list of parameters controlling the plot options: surv, risk, schoenfeld}

\item{Strat_Col}{column to stratify by if needed}

\item{time_lims}{limits for x axis of plot}

\item{age_unit}{age unit}

\item{model_control}{controls which alternative model options are used, see Def_model_control() for options and the Alternative_Regression_Options vignette for further details}
}
\value{
saves the plots in the current directory and returns a string that it passed
}
\description{
calculates and plots survival curves for each unique value of the stratification column
\code{CoxStratifiedSurvival} uses user provided data, columns, and identifier to calculate the survival fraction for each strata
}
\seealso{
Other {Plotting Functions}: 
\code{\link{CoxKaplanMeier}()},
\code{\link{CoxMartingale}()},
\code{\link{CoxRisk}()},
\code{\link{CoxSurvival}()},
\code{\link{GetCensWeight}()},
\code{\link{PlotCox_Schoenfeld_Residual}()}
}
\concept{{Plotting Functions}}
