% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTypes.R
\name{CoxSurvival}
\alias{CoxSurvival}
\title{calculates and plots survival plots of the estimated baseline}
\usage{
CoxSurvival(t, h, ch, surv, Plot_Name, verbose, time_lims, age_unit)
}
\arguments{
\item{t}{event times}

\item{h}{hazards of baseline}

\item{ch}{cumulative hazards of baseline}

\item{surv}{survival fraction of baseline}

\item{Plot_Name}{plot identifier, used in filename for saved plots}

\item{verbose}{boolean to control if additional information is printed to the console, also accepts 0/1 integer}

\item{time_lims}{limits for x axis of plot}

\item{age_unit}{age unit}
}
\value{
saves the plots in the current directory and returns a string that it passed
}
\description{
\code{CoxSurvival} uses user provided data, columns, and identifier to create plots
}
\seealso{
Other {Plotting Functions}: 
\code{\link{CoxKaplanMeier}()},
\code{\link{CoxMartingale}()},
\code{\link{CoxRisk}()},
\code{\link{CoxStratifiedSurvival}()},
\code{\link{GetCensWeight}()},
\code{\link{PlotCox_Schoenfeld_Residual}()}
}
\concept{{Plotting Functions}}
