% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Date_Shift}
\alias{Date_Shift}
\title{Automates creating a date difference column
\code{Date_Shift} generates a new dataframe with a column containing time difference in a given unit}
\usage{
Date_Shift(df, dcol0, dcol1, col_name, units = "days")
}
\arguments{
\item{df}{dataframe of data to use as reference}

\item{dcol0}{list of starting month, day, and year}

\item{dcol1}{list of ending month, day, and year}

\item{col_name}{new column name}

\item{units}{time unit to use}
}
\value{
returns the updated dataframe
}
\description{
Automates creating a date difference column
\code{Date_Shift} generates a new dataframe with a column containing time difference in a given unit
}
