% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Def_Control}
\alias{Def_Control}
\title{Automatically assigns missing control values}
\usage{
Def_Control(control)
}
\arguments{
\item{control}{list of parameters controlling the convergence, see the Control_Options vignette for details}
}
\value{
returns a filled list
}
\description{
\code{Def_Control} checks and assigns default values
}
\seealso{
Other Data Cleaning Functions: 
\code{\link{Check_Dupe_Columns}()},
\code{\link{Check_Trunc}()},
\code{\link{Check_Verbose}()},
\code{\link{Date_Shift}()},
\code{\link{Def_model_control}()},
\code{\link{Event_Count_Gen}()},
\code{\link{Event_Time_Gen}()},
\code{\link{Joint_Multiple_Events}()},
\code{\link{Replace_Missing}()},
\code{\link{Time_Since}()},
\code{\link{factorize}()},
\code{\link{gen_time_dep}()},
\code{\link{interact_them}()},
\code{\link{nested_split}()},
\code{\link{parse_literal_string}()}
}
\concept{Data Cleaning Functions}
