% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Def_Control}
\alias{Def_Control}
\title{Automatically assigns missing control values
\code{Def_Control} checks and assigns default values}
\usage{
Def_Control(control)
}
\arguments{
\item{control}{list of parameters controlling the convergence}
}
\value{
returns a filled list
}
\description{
Automatically assigns missing control values
\code{Def_Control} checks and assigns default values
}
\examples{
library(data.table)
control <- list("Ncores"=2,'lr' = 0.75,'maxiter' = 5, 'ties'='breslow','double_step'=1)
control <- Def_Control(control)

}
