% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Def_model_control}
\alias{Def_model_control}
\title{Automatically assigns missing model control values
\code{Def_model_control} checks and assigns default values}
\usage{
Def_model_control(control)
}
\arguments{
\item{control}{list of parameters controlling the convergence}
}
\value{
returns a filled list
}
\description{
Automatically assigns missing model control values
\code{Def_model_control} checks and assigns default values
}
\examples{
library(data.table)
control <- list("single"=TRUE)
control <- Def_model_control(control)

}
