% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Event_Time_Gen}
\alias{Event_Time_Gen}
\title{uses a table, list of categories, list of summaries, list of events, and person-year information to generate person-time tables}
\usage{
Event_Time_Gen(
  table,
  pyr = list(),
  time_scale = list(),
  categ = list(),
  summaries = list(),
  events = c(),
  fcount = FALSE,
  lcount = FALSE,
  studyid = "studyID",
  verbose = FALSE
)
}
\arguments{
\item{table}{dataframe with every category/event column needed}

\item{pyr}{list with entry and exit lists, containing day/month/year columns in the table}

\item{time_scale}{list with the time scale information, either a calendar category or an age category}

\item{categ}{list with category columns and methods, methods can be either strings or lists of boundaries}

\item{summaries}{list of columns to summarize, supports counts, means, and weighted means by person-year and renaming the summary column}

\item{events}{list of events or interests, checks if events are within each time interval}

\item{fcount}{boolean if the \verb{first at risk} count should be returned in the data. Returns the number of observations starting in each combination of categories.}

\item{lcount}{boolean if the \verb{last at risk} count should be returned in the data. Returns the number of observations ending in each combination of categories.}

\item{studyid}{id used to determine distinct subjects used for first and last at risk intervals}

\item{verbose}{integer valued 0-4 controlling what information is printed to the terminal. Each level includes the lower levels. 0: silent, 1: errors printed, 2: warnings printed, 3: notes printed, 4: debug information printed. Errors are situations that stop the regression, warnings are situations that assume default values that the user might not have intended, notes provide information on regression progress, and debug prints out C++ progress and intermediate results. The default level is 2 and True/False is converted to 3/0.}
}
\value{
returns a grouped table and a list of category boundaries used
}
\description{
\code{Event_Time_Gen} generates event-time tables
}
\examples{
library(data.table)
a <- c(0, 1, 2, 3, 4, 5, 6)
b <- c(1, 2, 3, 4, 5, 6, 7)
c <- c(0, 1, 0, 0, 0, 1, 0)
d <- c(1, 2, 3, 4, 5, 6, 7)
e <- c(2, 3, 4, 5, 6, 7, 8)
f <- c(
  1900, 1900, 1900, 1900,
  1900, 1900, 1900
)
g <- c(1, 2, 3, 4, 5, 6, 7)
h <- c(2, 3, 4, 5, 6, 7, 8)
i <- c(
  1901, 1902, 1903, 1904,
  1905, 1906, 1907
)
table <- data.table::data.table(
  a = a, b = b, c = c,
  d = d, e = e, f = f,
  g = g, h = h, i = i
)
categ <- list(
  a = "-1/3/5]7"
)
calendar_categ <- list(
  type = "calendar",
  day = c(1, 1, 1),
  month = c(1, 1, 1),
  year = c(1901, 1904, 1908)
)
time_scale <- list("time" = calendar_categ)
summary <- list(
  c = "count AS cases"
)
events <- list("c")
pyr <- list(
  entry = list(year = "f"),
  exit = list(year = "i"),
  unit = "years"
)
e <- Event_Time_Gen(table, pyr, time_scale, categ, summary, events)

}
\seealso{
Other Data Cleaning Functions:
\code{\link[=Date_Shift]{Date_Shift()}},
\code{\link[=Event_Count_Gen]{Event_Count_Gen()}},
\code{\link[=Joint_Multiple_Events]{Joint_Multiple_Events()}},
\code{\link[=Replace_Missing]{Replace_Missing()}},
\code{\link[=Time_Since]{Time_Since()}},
\code{\link[=factorize]{factorize()}},
\code{\link[=gen_time_dep]{gen_time_dep()}}
}
\concept{Data Cleaning Functions}
