% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableFunctions.R
\name{Event_Time_Gen}
\alias{Event_Time_Gen}
\title{uses a table, list of categories, list of summaries, list of events, and person-year information to generate person-time tables}
\usage{
Event_Time_Gen(
  table,
  pyr = list(),
  time_scale = list(),
  categ = list(),
  summaries = list(),
  events = c(),
  fcount = FALSE,
  lcount = FALSE,
  studyid = "studyID"
)
}
\arguments{
\item{table}{dataframe with every category/event column needed}

\item{pyr}{list with entry and exit lists, containing day/month/year columns in the table}

\item{time_scale}{list with the time scale information, either a calendar category or an age category}

\item{categ}{list with category columns and methods, methods can be either strings or lists of boundaries}

\item{summaries}{list of columns to summarize, supports counts, means, and weighted means by person-year and renaming the summary column}

\item{events}{list of events or interests, checks if events are within each time interval}

\item{fcount}{boolean if the \verb{first at risk} count should be returned in the data. Returns the number of observations starting in each combination of categories.}

\item{lcount}{boolean if the \verb{last at risk} count should be returned in the data. Returns the number of observations ending in each combination of categories.}

\item{studyid}{id used to determine distinct subjects used for first and last at risk intervals}
}
\value{
returns a grouped table and a list of category boundaries used
}
\description{
\code{Event_Time_Gen} generates event-time tables
}
\examples{
library(data.table)
a <- c(0, 1, 2, 3, 4, 5, 6)
b <- c(1, 2, 3, 4, 5, 6, 7)
c <- c(0, 1, 0, 0, 0, 1, 0)
d <- c(1, 2, 3, 4, 5, 6, 7)
e <- c(2, 3, 4, 5, 6, 7, 8)
f <- c(
  1900, 1900, 1900, 1900,
  1900, 1900, 1900
)
g <- c(1, 2, 3, 4, 5, 6, 7)
h <- c(2, 3, 4, 5, 6, 7, 8)
i <- c(
  1901, 1902, 1903, 1904,
  1905, 1906, 1907
)
table <- data.table::data.table(
  a = a, b = b, c = c,
  d = d, e = e, f = f,
  g = g, h = h, i = i
)
categ <- list(
  a = "-1/3/5]7"
)
calendar_categ <- list(
  type = "calendar",
  day = c(1, 1, 1),
  month = c(1, 1, 1),
  year = c(1901, 1904, 1908)
)
time_scale <- list("time" = calendar_categ)
summary <- list(
  c = "count AS cases"
)
events <- list("c")
pyr <- list(
  entry = list(year = "f"),
  exit = list(year = "i"),
  unit = "years"
)
e <- Event_Time_Gen(table, pyr, time_scale, categ, summary, events)

}
\seealso{
Other Table Generation Functions:
\code{\link[=Event_Count_Gen]{Event_Count_Gen()}}
}
\concept{Table Generation Functions}
