% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{EventAssignment.default}
\alias{EventAssignment.default}
\title{Predicts how many events are due to baseline vs excess}
\usage{
\method{EventAssignment}{default}(x, df, ...)
}
\arguments{
\item{x}{result object from a regression, class poisres}

\item{df}{a data.table containing the columns of interest}

\item{...}{extended for other necessary parameters}
}
\description{
\code{EventAssignment} Generic background/excess event calculation function, by default nothing happens
}
