% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{EventAssignment}
\alias{EventAssignment}
\title{Generic background/excess event calculation function}
\usage{
EventAssignment(x, df, ...)
}
\arguments{
\item{x}{result object from a regression, class poisres}

\item{df}{a data.table containing the columns of interest}

\item{...}{extended for other necessary parameters}
}
\description{
\code{EventAssignment} Generic background/excess event calculation function for
standard poisson regressions and likelihood based confidence interval results.
}
\seealso{
Other Poisson Event Assignment:
\code{\link[=EventAssignment.poisres]{EventAssignment.poisres()}},
\code{\link[=EventAssignment.poisresbound]{EventAssignment.poisresbound()}}
}
\concept{Poisson Event Assignment}
