% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{factorize}
\alias{factorize}
\title{Splits a parameter into factors
\code{factorize} uses user provided list of columns to define new parameter for each unique value and update the data.table.
Not for interaction terms}
\usage{
factorize(df, col_list)
}
\arguments{
\item{df}{a data.table containing the columns of interest}

\item{col_list}{an array of column names that should have factor terms defined}
}
\value{
returns a list with two named fields. df for the updated dataframe, and cols for the new column names
}
\description{
Splits a parameter into factors
\code{factorize} uses user provided list of columns to define new parameter for each unique value and update the data.table.
Not for interaction terms
}
