% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{gen_time_dep}
\alias{gen_time_dep}
\title{Applies time dependence to parameters
\code{gen_time_dep} generates a new dataframe with time dependent covariates by applying a grid in time}
\usage{
gen_time_dep(
  df,
  time0,
  time1,
  event,
  iscox,
  dt,
  new_names,
  dep_cols,
  func_form,
  fname,
  tform
)
}
\arguments{
\item{df}{dataframe of data to use as reference}

\item{time0}{starting time column}

\item{time1}{ending time column}

\item{event}{event column}

\item{iscox}{boolean if rows not at event times should be kept}

\item{dt}{spacing in time for new rows}

\item{new_names}{vector of new column names for the time dependent columns}

\item{dep_cols}{columns that are not needed in the new dataframe}

\item{func_form}{vector of functions to apply to each time-dependent covariate. Of the form func(df, time) returning a vector of the new column value}

\item{fname}{filename used for new dataframe}

\item{tform}{list of string function identifiers, used for linear/step}
}
\value{
returns the updated dataframe
}
\description{
Applies time dependence to parameters
\code{gen_time_dep} generates a new dataframe with time dependent covariates by applying a grid in time
}
