% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvTypes.R
\name{get_form_risk}
\alias{get_form_risk}
\title{Interprets the risk factor side of a formula}
\usage{
get_form_risk(model_obj, df)
}
\arguments{
\item{model_obj}{output from get_form_risk, list of risk factor model values}

\item{df}{a data.table containing the columns of interest}
}
\value{
returns the right hand side of a formula reading
}
\description{
\code{get_form_risk} interprets the RHS of a formula
}
\seealso{
Other Formula Interpretation: 
\code{\link{ColossusCaseConStrataSurv}()},
\code{\link{ColossusCaseConSurv}()},
\code{\link{ColossusCaseConTimeStrataSurv}()},
\code{\link{ColossusCaseConTimeSurv}()},
\code{\link{ColossusCoxStrataSurv}()},
\code{\link{ColossusCoxSurv}()},
\code{\link{ColossusExpressionCall}()},
\code{\link{ColossusFineGrayStrataSurv}()},
\code{\link{ColossusFineGraySurv}()},
\code{\link{ColossusPoisSurv}()},
\code{\link{get_form}()},
\code{\link{get_form_joint}()},
\code{\link{get_form_list}()},
\code{\link{get_form_surv}()}
}
\concept{Formula Interpretation}
