% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Likelihood_Ratio_Test}
\alias{Likelihood_Ratio_Test}
\title{Defines the likelihood ratio test
\code{Likelihood_Ratio_Test} uses two models and calculates the ratio}
\usage{
Likelihood_Ratio_Test(alternative_model, null_model)
}
\arguments{
\item{alternative_model}{the new model of interest}

\item{null_model}{a model to compare against}
}
\value{
returns the score statistic
}
\description{
Defines the likelihood ratio test
\code{Likelihood_Ratio_Test} uses two models and calculates the ratio
}
