% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{LikelihoodBound.logitres}
\alias{LikelihoodBound.logitres}
\title{Calculates the likelihood boundary for a completed Logistic model}
\usage{
\method{LikelihoodBound}{logitres}(x, df, curve_control = list(), control = list(), ...)
}
\arguments{
\item{x}{result object from a regression, class poisres}

\item{df}{a data.table containing the columns of interest}

\item{curve_control}{a list of control options for the likelihood boundary regression. See the \href{../articles/Control_Options.html}{control options vignette} for details.}

\item{control}{list of parameters controlling the convergence, see the \href{../articles/Control_Options.html}{control options vignette} for details}

\item{...}{can include the named entries for the curve_control list parameter}
}
\value{
returns a list of the final results
}
\description{
\code{LikelihoodBound.logitres} solves the confidence interval for a Poisson model, starting at the optimum point and
iteratively optimizing end-points of intervals.
}
\seealso{
Other Logistic Wrapper Functions:
\code{\link[=LogisticRun]{LogisticRun()}},
\code{\link[=LogisticRunMulti]{LogisticRunMulti()}}
}
\concept{Likelihood Boundaries}
