% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{LikelihoodBound.poisres}
\alias{LikelihoodBound.poisres}
\title{Calculates the likelihood boundary for a completed Poisson model}
\usage{
\method{LikelihoodBound}{poisres}(object, df, curve_control = list(), control = list(), ...)
}
\arguments{
\item{df}{a data.table containing the columns of interest}

\item{control}{list of parameters controlling the convergence, see Def_Control() for options or vignette("Control_Options")}
}
\value{
returns a list of the final results
}
\description{
\code{LikelihoodBound.poisres} solves the confidence interval for a Poisson model, starting at the optimum point and
iteratively optimizing end-points of intervals.
}
\seealso{
Other Cox Wrapper Functions: 
\code{\link{CoxCurveSolver}()},
\code{\link{CoxRun}()},
\code{\link{CoxRunMulti}()},
\code{\link{LikelihoodBound.coxres}()},
\code{\link{RunCaseControlRegression_Omnibus}()},
\code{\link{RunCoxRegression_Omnibus}()},
\code{\link{RunCoxRegression_Omnibus_Multidose}()}
}
\concept{Cox Wrapper Functions}
