% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{LikelihoodBound}
\alias{LikelihoodBound}
\title{Generic likelihood boundary calculation function}
\usage{
LikelihoodBound(x, df, curve_control = list(), control = list(), ...)
}
\arguments{
\item{x}{result object from a regression, class coxres or poisres}

\item{df}{a data.table containing the columns of interest}

\item{curve_control}{a list of control options for the likelihood boundary regression. See the Control_Options vignette for details.}

\item{control}{list of parameters controlling the convergence, see the Control_Options vignette for details}

\item{...}{extended for other necessary parameters}
}
\description{
\code{LikelihoodBound} Generic likelihood boundary calculation function
}
