% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Linked_Dose_Formula}
\alias{Linked_Dose_Formula}
\title{Calculates Full Parameter list for Special Dose Formula
\code{Linked_Dose_Formula} Calculates all parameters for linear-quadratic and linear-exponential linked formulas}
\usage{
Linked_Dose_Formula(tforms, paras, verbose = FALSE)
}
\arguments{
\item{tforms}{list of formula types}

\item{paras}{list of formula parameters}

\item{verbose}{verbosity argument, for error returns}
}
\value{
returns list of full parameters
}
\description{
Calculates Full Parameter list for Special Dose Formula
\code{Linked_Dose_Formula} Calculates all parameters for linear-quadratic and linear-exponential linked formulas
}
