% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Linked_Lin_Exp_Para}
\alias{Linked_Lin_Exp_Para}
\title{Calculates The Additional Parameter For a linear-exponential formula with known maximum
\code{Linked_Lin_Exp_Para} Calculates what the additional parameter would be for a desired maximum}
\usage{
Linked_Lin_Exp_Para(y, a0, a1_goal)
}
\arguments{
\item{y}{point formula switch}

\item{a0}{linear slope}

\item{a1_goal}{exponential maximum desired}
}
\value{
returns parameter used by Colossus
}
\description{
Calculates The Additional Parameter For a linear-exponential formula with known maximum
\code{Linked_Lin_Exp_Para} Calculates what the additional parameter would be for a desired maximum
}
