% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Make_subterms_Basic}
\alias{Make_subterms_Basic}
\title{Utility function to calculate the term and subterm values with the basic model
\code{Make_subterms_Basic} Called to update term matrices, Uses lists of term numbers and types to apply formulas}
\usage{
Make_subterms_Basic(totalnum, dfc, T0, beta_0, df0, nthreads, debugging)
}
\arguments{
\item{totalnum}{total number of parameters}

\item{dfc}{vector matching subterm number to matrix column}

\item{T0}{Term value for each subterm}

\item{beta_0}{parameter estimates}

\item{df0}{matrix of covariate values}

\item{nthreads}{number of threads available}

\item{debugging}{additional boolean for verbosity in testing}
}
\value{
Updates matrices in place: subterm matrices, Term matrices
}
\description{
Utility function to calculate the term and subterm values with the basic model
\code{Make_subterms_Basic} Called to update term matrices, Uses lists of term numbers and types to apply formulas
}
