% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{nested_split}
\alias{nested_split}
\title{Splits a string by commas, skipping over parenthesis sections}
\usage{
nested_split(total_string)
}
\arguments{
\item{total_string}{the complete string to split}
}
\value{
returns a vector of substrings
}
\description{
\code{nested_split} splits by comma, keeps parenthesis sections together
}
\seealso{
Other Data Cleaning Functions: 
\code{\link{Check_Dupe_Columns}()},
\code{\link{Check_Trunc}()},
\code{\link{Check_Verbose}()},
\code{\link{Date_Shift}()},
\code{\link{Def_Control}()},
\code{\link{Def_model_control}()},
\code{\link{Event_Count_Gen}()},
\code{\link{Event_Time_Gen}()},
\code{\link{Joint_Multiple_Events}()},
\code{\link{Replace_Missing}()},
\code{\link{Time_Since}()},
\code{\link{factorize}()},
\code{\link{gen_time_dep}()},
\code{\link{interact_them}()},
\code{\link{parse_literal_string}()}
}
\concept{Data Cleaning Functions}
