% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTypes.r
\name{PlotCox_Schoenfeld_Residual}
\alias{PlotCox_Schoenfeld_Residual}
\title{Calculates Schoenfeld residuals for a Cox Proportional Hazards regression and plots
\code{RunCox_Schoenfeld_Residual} uses user provided data, time/event columns, vectors specifying the model, and options to calculate the residuals}
\usage{
PlotCox_Schoenfeld_Residual(
  df,
  time1,
  time2,
  event,
  names,
  Term_n,
  tform,
  keep_constant,
  a_n,
  modelform,
  fir,
  der_iden,
  control,
  age_unit,
  Plot_Name,
  alpha = 0.05
)
}
\arguments{
\item{df}{data used for regression}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{event}{column used for event status}

\item{names}{columns names for elements of the model, used to identify data columns}

\item{Term_n}{term numbers for each element of the model}

\item{tform}{subterm type for each element of the model}

\item{keep_constant}{vector of 0/1 to identify parameters to force to be constant}

\item{a_n}{starting parameters for regression}

\item{modelform}{string specifying the model type}

\item{fir}{term number for the intial term, used for models of the form T0*f(Ti) in which the order matters}

\item{der_iden}{number for the subterm to test derivative at, only used for testing runs with a single varying parameter}

\item{control}{list of parameters controlling the convergence}

\item{age_unit}{age unit}

\item{Plot_Name}{plot identifier}

\item{alpha}{significance level for two tail t test}
}
\value{
returns a list of the final results
}
\description{
Calculates Schoenfeld residuals for a Cox Proportional Hazards regression and plots
\code{RunCox_Schoenfeld_Residual} uses user provided data, time/event columns, vectors specifying the model, and options to calculate the residuals
}
