% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{plotRisk}
\alias{plotRisk}
\title{Generic Risk Plotting function}
\usage{
plotRisk(x, df, plot_options, a_n = c(), ...)
}
\arguments{
\item{x}{result object from a regression, class coxres}

\item{df}{a data.table containing the columns of interest}

\item{plot_options}{list of parameters controlling the plot options, see the \href{../articles/Plotting_And_Analysis.html}{plotting vignette} for different options}

\item{a_n}{list of initial parameter values, used to determine the number of parameters. May be either a list of vectors or a single vector.}

\item{...}{can include the named entries for the plot_options parameter}
}
\description{
\code{plotRisk} Generic risk plotting function. Used by the coxres
object to plot the hazard ratios at different levels of individual
covariates. Currently only returns data that could be plotted.
}
