% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{plotSchoenfeld}
\alias{plotSchoenfeld}
\title{Generic Schoenfeld Residual Plotting function}
\usage{
plotSchoenfeld(x, df, plot_options, a_n = c(), ...)
}
\arguments{
\item{x}{result object from a regression, class coxres}

\item{df}{a data.table containing the columns of interest}

\item{plot_options}{list of parameters controlling the plot options, see \code{RunCoxPlots()} for different options}

\item{a_n}{list of initial parameter values, used to determine the number of parameters. May be either a list of vectors or a single vector.}

\item{...}{can include the named entries for the plot_options parameter}
}
\description{
\code{plotSchoenfeld} Generic Schoenfeld Residual Plotting
}
