% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{PoisRunJoint}
\alias{PoisRunJoint}
\title{Fully runs a joint poisson regression model, returning the model and results}
\usage{
PoisRunJoint(
  model,
  table,
  a_n = list(c(0)),
  keep_constant = c(0),
  control = list(),
  gradient_control = list(),
  single = FALSE,
  oberved_info = FALSE,
  cons_mat = as.matrix(c(0)),
  cons_vec = c(0),
  ...
)
}
\arguments{
\item{a_n}{list of initial parameter values, used to determine the number of parameters. May be either a list of vectors or a single vector.}

\item{keep_constant}{binary values to denote which parameters to change}

\item{control}{list of parameters controlling the convergence, see Def_Control() for options or vignette("Control_Options")}

\item{cons_mat}{Matrix containing coefficients for a system of linear constraints, formatted as matrix}

\item{cons_vec}{Vector containing constants for a system of linear constraints, formatted as vector}
}
\value{
returns a class fully describing the model and the regression results
}
\description{
\code{PoisRunJoint} uses a list of formula, data.table, and list of controls to prepare and
run a Colossus poisson regression function on a joint dataset
}
\examples{
library(data.table)
df <- data.table::data.table(
  "UserID" = c(112, 114, 213, 214, 115, 116, 117),
  "Starting_Age" = c(18, 20, 18, 19, 21, 20, 18),
  "Ending_Age" = c(30, 45, 57, 47, 36, 60, 55),
  "Cancer_Status" = c(0, 0, 1, 0, 1, 0, 0),
  "Flu_Status"    = c(0, 1, 0, 0, 1, 0, 1),
  "a" = c(0, 1, 1, 0, 1, 0, 1),
  "b" = c(1, 1.1, 2.1, 2, 0.1, 1, 0.2),
  "c" = c(10, 11, 10, 11, 12, 9, 11),
  "d" = c(0, 0, 0, 1, 1, 1, 1),
  "e" = c(0, 0, 1, 0, 0, 0, 1)
)
formula_list <- list(Pois(Ending_Age, Cancer_Status) ~ plinear(d, 0),
                     Pois(Ending_Age, Flu_Status)    ~ loglinear(d, 0),
                 "shared" = Pois(Ending_Age)    ~ loglinear(a, b, c, 0)
)
res <- PoisRunJoint(formula_list, df)
}
\seealso{
Other Poisson Wrapper Functions: 
\code{\link{PoisRun}()},
\code{\link{PoissonCurveSolver}()},
\code{\link{RunPoissonEventAssignment}()},
\code{\link{RunPoissonEventAssignment_bound}()},
\code{\link{RunPoissonRegression_Joint_Omnibus}()},
\code{\link{RunPoissonRegression_Omnibus}()},
\code{\link{RunPoissonRegression_Residual}()}
}
\concept{Poisson Wrapper Functions}
