% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{print.logitres}
\alias{print.logitres}
\title{Prints a logistic regression output clearly}
\usage{
\method{print}{logitres}(x, ...)
}
\arguments{
\item{x}{result object from a regression, class logitres}

\item{...}{can include the number of digits, named digit, or an unnamed integer entry assumed to be digits}
}
\value{
return nothing, prints the results to console
}
\description{
\code{print.logitres} uses the list output from a regression, prints off a table of results and summarizes the score and convergence.
}
\seealso{
Other Output and Information Functions: 
\code{\link{System_Version}()},
\code{\link{print.caseconres}()},
\code{\link{print.coxres}()},
\code{\link{print.coxresbound}()},
\code{\link{print.poisres}()},
\code{\link{print.poisresbound}()}
}
\concept{Output and Information Functions}
