% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{RelativeRisk}
\alias{RelativeRisk}
\title{Generic relative risk calculation function}
\usage{
RelativeRisk(x, df, ...)
}
\arguments{
\item{x}{result object from a regression, class coxres}

\item{df}{a data.table containing the columns of interest}

\item{...}{extended for other necessary parameters}
}
\description{
\code{RelativeRisk} Generic relative risk calculation function,
currently only used by the coxres object, to calculate relative risks
for either the last parameter or a custom point.
}
\seealso{
Other Predicted Risk and Rate:
\code{\link[=RelativeRisk.coxres]{RelativeRisk.coxres()}}
}
\concept{Predicted Risk and Rate}
