% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Replace_Missing}
\alias{Replace_Missing}
\title{Automatically assigns missing values in listed columns
\code{Replace_Missing} checks each column and fills in NA values}
\usage{
Replace_Missing(df, name_list, MSV, verbose = FALSE)
}
\arguments{
\item{df}{datatable being replaced in}

\item{name_list}{vector of string column names to check}

\item{MSV}{value to replace na with, same used for every column used}

\item{verbose}{verbosity for printing which columns needed fills}
}
\value{
returns a filled datatable
}
\description{
Automatically assigns missing values in listed columns
\code{Replace_Missing} checks each column and fills in NA values
}
