% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{Residual.logitmodel}
\alias{Residual.logitmodel}
\title{Calculates the Residuals for a logistic model}
\usage{
\method{Residual}{logitmodel}(
  x,
  df,
  control = list(),
  a_n = NULL,
  link = "odds",
  pearson = FALSE,
  deviance = FALSE,
  ...
)
}
\arguments{
\item{x}{result model for a regression, class logitmodel}

\item{df}{a data.table containing the columns of interest}

\item{control}{list of parameters controlling the convergence, see the \href{../articles/Control_Options.html}{control options vignette} for details}

\item{a_n}{list of initial parameter values, used to determine the number of parameters. May be either a list of vectors or a single vector.}

\item{link}{Used in logistic regression, the linking function relating the input model and event probability. Current options are "odds", "ident", and "loglink" for the odds ratio, identity, and complimentary loglink options.}

\item{pearson}{boolean to calculate pearson residuals}

\item{deviance}{boolean to calculate deviance residuals}

\item{...}{can include the named entries for the assign_control list parameter}
}
\value{
returns a list of the final results
}
\description{
\code{Residual.logitmodel} uses user provided data, person-year/event columns, vectors specifying the model,
and options to calculate residuals for a solved Poisson regression
}
\seealso{
Other Residuals:
\code{\link[=Residual]{Residual()}},
\code{\link[=Residual.logitres]{Residual.logitres()}},
\code{\link[=Residual.poismodel]{Residual.poismodel()}},
\code{\link[=Residual.poisres]{Residual.poisres()}}
}
\concept{Residuals}
