% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{Residual.poisres}
\alias{Residual.poisres}
\title{Calculates the likelihood boundary for a completed cox model}
\usage{
\method{Residual}{poisres}(
  x,
  df,
  control = list(),
  a_n = c(),
  pearson = FALSE,
  deviance = FALSE,
  ...
)
}
\arguments{
\item{x}{result object from a regression, class poisres
#param pearson boolean to calculate pearson residuals
#param deviance boolean to calculate deviance residuals}

\item{df}{a data.table containing the columns of interest}

\item{control}{list of parameters controlling the convergence, see the Control_Options vignette for details}

\item{a_n}{list of initial parameter values, used to determine the number of parameters. May be either a list of vectors or a single vector.}

\item{...}{can include the named entries for the assign_control list parameter}
}
\value{
returns a list of the final results
}
\description{
\code{LikelihoodBound.coxres} uses user provided data, person-year/event columns, vectors specifying the model,
and options to calculate background and excess events for a solved Poisson regression
}
\seealso{
Other Poisson Wrapper Functions: 
\code{\link{EventAssignment.poisres}()},
\code{\link{LikelihoodBound.poisres}()},
\code{\link{PoisRun}()},
\code{\link{PoisRunJoint}()},
\code{\link{PoissonCurveSolver}()},
\code{\link{RunPoissonEventAssignment}()},
\code{\link{RunPoissonEventAssignment_bound}()},
\code{\link{RunPoissonRegression_Joint_Omnibus}()},
\code{\link{RunPoissonRegression_Omnibus}()},
\code{\link{RunPoissonRegression_Residual}()}
}
\concept{Poisson Wrapper Functions}
