% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicRun.R
\name{Residual}
\alias{Residual}
\title{Generic Residual calculation function}
\usage{
Residual(x, df, ...)
}
\arguments{
\item{x}{result object from a regression, class coxres or poisres}

\item{df}{a data.table containing the columns of interest}

\item{...}{extended for other necessary parameters}
}
\description{
\code{Residual} Generic residual calculation function for regression
results. Note that the Cox residuals are calculated using the various
plotting based functions.
}
\seealso{
Other Residuals:
\code{\link[=Residual.logitmodel]{Residual.logitmodel()}},
\code{\link[=Residual.logitres]{Residual.logitres()}},
\code{\link[=Residual.poismodel]{Residual.poismodel()}},
\code{\link[=Residual.poisres]{Residual.poisres()}}
}
\concept{Residuals}
