% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{risk_check_transition}
\alias{risk_check_transition}
\title{Interface between R code and the risk check
\code{cox_ph_transition_single} Called directly from R, Defines the control variables and calls the function which only calculates the log-likelihood}
\usage{
risk_check_transition(
  Term_n,
  tform,
  a_n,
  dfc,
  x_all,
  fir,
  modelform,
  Control,
  model_control,
  KeepConstant,
  term_tot
)
}
\arguments{
\item{Term_n}{Term numbers}

\item{tform}{subterm types}

\item{a_n}{starting values}

\item{dfc}{vector matching subterm number to matrix column}

\item{x_all}{covariate matrix}

\item{fir}{st term number}

\item{modelform}{string model identifier}

\item{Control}{control list}

\item{model_control}{controls which alternative model options are used}

\item{KeepConstant}{binary vector to identify constant parameters}

\item{term_tot}{total number of terms}
}
\value{
LogLik_Cox_PH output : Log-likelihood of optimum, first derivative of log-likelihood, second derivative matrix, parameter list, standard deviation estimate, AIC, model information
}
\description{
Interface between R code and the risk check
\code{cox_ph_transition_single} Called directly from R, Defines the control variables and calls the function which only calculates the log-likelihood
}
