% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cox_Regression.R
\name{RunCoxNull}
\alias{RunCoxNull}
\title{Performs basic Cox Proportional Hazards regression with the null model
\code{RunCoxRegression} uses user provided data and time/event columns to calculate the log-likelihood with constant hazard ratio}
\usage{
RunCoxNull(
  df,
  time1 = "age_start",
  time2 = "age_exit",
  event0 = "cases",
  control = list(lr = 0.75, maxiter = 20, halfmax = 5, epsilon = 1e-09, dbeta_max = 0.5,
    deriv_epsilon = 1e-09, abs_max = 1, change_all = TRUE, dose_abs_max = 100, verbose =
    FALSE, ties = "breslow", double_step = 1)
)
}
\arguments{
\item{df}{data used for regression}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{event0}{column used for event status}

\item{control}{list of parameters controlling the convergence}
}
\value{
returns a list of the final results
}
\description{
Performs basic Cox Proportional Hazards regression with the null model
\code{RunCoxRegression} uses user provided data and time/event columns to calculate the log-likelihood with constant hazard ratio
}
