% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cox_Regression.R
\name{RunCoxPlots}
\alias{RunCoxPlots}
\title{Performs Cox Proportional Hazard model plots
\code{RunCoxPlots} uses user provided data, time/event columns, vectors specifying the model, and options to choose plots and saves plots}
\usage{
RunCoxPlots(
  df,
  time1,
  time2,
  event0,
  names,
  Term_n,
  tform,
  keep_constant,
  a_n,
  modelform,
  fir,
  control,
  plot_options
)
}
\arguments{
\item{df}{data used for regression}

\item{time1}{column used for time period starts}

\item{time2}{column used for time period end}

\item{event0}{column used for event status}

\item{names}{columns names for elements of the model, used to identify data columns}

\item{Term_n}{term numbers for each element of the model}

\item{tform}{subterm type for each element of the model}

\item{keep_constant}{vector of 0/1 to identify parameters to force to be constant}

\item{a_n}{starting parameters for regression}

\item{modelform}{string specifying the model type}

\item{fir}{term number for the initial term, used for models of the form T0*f(Ti) in which the order matters}

\item{control}{list of parameters controlling the convergence}

\item{plot_options}{list of parameters controlling the plot options}
}
\value{
saves the plots in the current directory and returns a string that it passed
}
\description{
Performs Cox Proportional Hazard model plots
\code{RunCoxPlots} uses user provided data, time/event columns, vectors specifying the model, and options to choose plots and saves plots
}
