% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_Regression.R
\name{RunPoissonRegression_Single}
\alias{RunPoissonRegression_Single}
\title{Performs basic poisson calculation
\code{RunPoissonRegression_Single} uses user provided data, person-year/event columns, vectors specifying the model, and returns the score}
\usage{
RunPoissonRegression_Single(
  df,
  pyr,
  event,
  names,
  Term_n,
  tform,
  a_n,
  modelform,
  fir,
  control
)
}
\arguments{
\item{df}{data used for regression}

\item{pyr}{column used for person-years per row}

\item{event}{column used for event status}

\item{names}{columns names for elements of the model, used to identify data columns}

\item{Term_n}{term numbers for each element of the model}

\item{tform}{subterm type for each element of the model}

\item{a_n}{starting parameters for regression}

\item{modelform}{string specifying the model type}

\item{fir}{term number for the intial term, used for models of the form T0*f(Ti) in which the order matters}

\item{control}{list of parameters controlling the convergence}
}
\value{
prints the final results and return null
}
\description{
Performs basic poisson calculation
\code{RunPoissonRegression_Single} uses user provided data, person-year/event columns, vectors specifying the model, and returns the score
}
