% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{System_Version}
\alias{System_Version}
\title{Checks OS, compilers, and OMP}
\usage{
System_Version()
}
\value{
returns a list of results
}
\description{
\code{System_Version} checks OS, default R c++ compiler, and if OMP is enabled
}
\seealso{
Other Output and Information Functions:
\code{\link[=print.caseconres]{print.caseconres()}},
\code{\link[=print.coxres]{print.coxres()}},
\code{\link[=print.coxresbound]{print.coxresbound()}},
\code{\link[=print.logitres]{print.logitres()}},
\code{\link[=print.logitresbound]{print.logitresbound()}},
\code{\link[=print.poisres]{print.poisres()}},
\code{\link[=print.poisresbound]{print.poisresbound()}}
}
\concept{Output and Information Functions}
