% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Time_Since}
\alias{Time_Since}
\title{Automates creating a date since a reference column
\code{Time_Since} generates a new dataframe with a column containing time since a reference in a given unit}
\usage{
Time_Since(df, dcol0, tref, col_name, units = "days")
}
\arguments{
\item{df}{dataframe of data to use as reference}

\item{dcol0}{list of starting month, day, and year}

\item{tref}{reference time in date format}

\item{col_name}{new column name}

\item{units}{time unit to use}
}
\value{
returns the updated dataframe
}
\description{
Automates creating a date since a reference column
\code{Time_Since} generates a new dataframe with a column containing time since a reference in a given unit
}
